% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_age_diff_gt_one}
\alias{chk_age_diff_gt_one}
\alias{err_age_diff_gt_one}
\title{Check that difference between upper and lower limits of
age groups greater than 1}
\usage{
chk_age_diff_gt_one(age_low, age_up, is_low_up, labels)

err_age_diff_gt_one(age_low, age_up, is_low_up, labels)
}
\arguments{
\item{age_low}{A numeric vector. The lower
breaks for the age groups.}

\item{age_up}{A numeric vector. The upper
breaks for the age groups.}

\item{is_low_up}{Logical. Whether label
has "low-up" format, eg "15-19".}

\item{labels}{A vector of age group labels.}
}
\description{
Check applies to age groups with labels that have
a "low-up" format, ie not single-year age groups,
not open age groups, and not missing.
}
\examples{
chk_age_diff_gt_one(age_low = c(0, 1, 5, 10),
                    age_up = c(1, 5, 10, NA),
                    is_low_up = c(TRUE, TRUE, TRUE, FALSE),
                    labels = c("0", "1-4", "5-9", "10+"))
}
