% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_age_ge_break_min}
\alias{chk_age_ge_break_min}
\alias{chk_age_lt_break_max}
\alias{err_age_ge_break_min}
\alias{err_age_lt_break_max}
\title{Check that all age groups fall within limits
imposed by 'break_min' or 'break_max'}
\usage{
chk_age_ge_break_min(labels, lower, break_min)

chk_age_lt_break_max(labels, upper, break_max)

err_age_ge_break_min(labels, lower, break_min)

err_age_lt_break_max(labels, upper, break_max)
}
\arguments{
\item{labels}{A vector of age group labels.}

\item{lower}{A numeric vector. The lower
breaks for the age groups.}

\item{break_min}{Integer. Ages must be equal
to or greater than this.
for \code{age}.}

\item{upper}{A numeric vector. The upper
breaks for the age groups.}

\item{break_max}{Integer. Ages must be
lower than this.}
}
\description{
All age groups are assumed to be closed.
}
\examples{
chk_age_ge_break_min(labels = c("15-19", "20", "100-104"),
                     lower = c(15, 20, 100, NA),
                     break_min = 15)
chk_age_lt_break_max(labels = c("20-24", "45-49", NA),
                     upper = c(25, 50, NA),
                     break_max = 50)
}
