% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_all_0_1}
\alias{chk_all_0_1}
\alias{err_all_0_1}
\title{Check that a vector consists entirely of of 1s and 0s}
\usage{
chk_all_0_1(x, name)

err_all_0_1(x, name)
}
\arguments{
\item{x}{An integer vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector consists entirely of of 1s and 0s
}
\examples{
x <- c(0L, 1L, 0L, 0L, 1L)
chk_all_0_1(x, name = "x")
}
