% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_all_x1_in_x2}
\alias{chk_all_x1_in_x2}
\alias{err_all_x1_in_x2}
\title{Check whether all elements of 'x1' are in 'x2'}
\usage{
chk_all_x1_in_x2(x1, x2, name1, name2, exclude_zero)

err_all_x1_in_x2(x1, x2, name1, name2, exclude_zero)
}
\arguments{
\item{x1}{A vector, typically integers.}

\item{x2}{A vector, typically integers.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}

\item{exclude_zero}{Whether to exclude zeros
from \code{x1} before testing.}
}
\description{
Check \code{x1} is contained in \code{x2}, potentially
after excluding \code{0}s from \code{x1}.
}
\examples{
x1 <- c(0L, 1L, 2L, 0L)
x2 <- 1:3
chk_all_x1_in_x2(x1 = x2,
                 x2 = x2,
                 name1 = "x1",
                 name2 = "x2",
                 exclude_zero = TRUE)
}
