% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_at_most_one_na_vector}
\alias{chk_at_most_one_na_vector}
\alias{chk_at_most_one_na_list}
\alias{err_at_most_one_na_vector}
\alias{err_at_most_one_na_list}
\title{Check that a vector or list has 0 or 1 NAs}
\usage{
chk_at_most_one_na_vector(x, name)

chk_at_most_one_na_list(x, name)

err_at_most_one_na_vector(x, name)

err_at_most_one_na_list(x, name)
}
\arguments{
\item{x}{An integer vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
In the case of a list, an item is 'NA'
if it has at least one element,
and of these elements are NA
}
\examples{
x <- c(0L, 1L, NA)
chk_at_most_one_na_vector(x, name = "x")
x <- c(c(0L, 1L), c(2L, NA))
chk_at_most_one_na_list(x, name = "x")
}
