% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_character_complete}
\alias{chk_character_complete}
\alias{err_character_complete}
\title{Check that a character vector has no
NAs, blanks, or duplicates}
\usage{
chk_character_complete(x, name)

err_character_complete(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a character vector has no
NAs, blanks, or duplicates
}
\examples{
x <- c("A", "B", "C")
chk_character_complete(x, name = "x")
}
