% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_comp_type_indices}
\alias{chk_comp_type_indices}
\alias{err_comp_type_indices}
\title{Check that 'i_comp_type_self' consistent with
'indices_orig_self' and 'i_direction_self'}
\usage{
chk_comp_type_indices(i_comp_type_self, indices_orig_self, i_direction_self)

err_comp_type_indices(i_comp_type_self, indices_orig_self, i_direction_self)
}
\arguments{
\item{i_comp_type_self}{Integer between 1 and 4}

\item{indices_orig_self}{Integer vector}

\item{i_direction_self}{Integer scalar}
}
\description{
Check that 'i_comp_type_self' consistent with
'indices_orig_self' and 'i_direction_self'
}
\examples{
chk_comp_type_indices(i_comp_type_self = 3L,
                      indices_orig_self = 2L,
                      i_direction_self = 0L)
}
