% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_dimnames_complete}
\alias{chk_dimnames_complete}
\alias{err_dimnames_complete}
\title{Check that all dimensions of an array have complete dimnames}
\usage{
chk_dimnames_complete(x, name)

err_dimnames_complete(x, name)
}
\arguments{
\item{x}{An array.}

\item{name}{The name for \code{x} that
will be displayed in error messages.}
}
\description{
Check that the dimnames for an array include every
dimension, and have no blanks, NAs, or duplicates
}
\examples{
x <- array(1:6,
           dim = 3:2,
           dimnames = list(region = c("a", "b", "c"),
                           sex = c("Female", "Male")))
chk_dimnames_complete(x, name = "x")
}
\seealso{
\code{\link{chk_array_metadata_complete}}.
}
