% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_dimtypes_mutually_compatible}
\alias{chk_dimtypes_mutually_compatible}
\alias{err_dimtypes_mutually_compatible}
\title{Check that dimtypes obey rules about appearing at most once,
or not occurring with other dimtypes}
\usage{
chk_dimtypes_mutually_compatible(dimtypes)

err_dimtypes_mutually_compatible(dimtypes)
}
\arguments{
\item{dimtypes}{A character vector of dimtypes.}
}
\description{
Check that dimtypes obey rules about appearing at most once,
or not occurring with other dimtypes
}
\examples{
dimtypes <- c("age", "time", "triangle")
chk_dimtypes_mutually_compatible(dimtypes)
}
\seealso{
\code{\link{chk_names_pairs_complete}}
}
