% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_finite_scalar}
\alias{chk_finite_scalar}
\alias{chk_finite_vector}
\alias{err_finite_scalar}
\alias{err_finite_vector}
\title{Check that a scalar or vector only has finite values}
\usage{
chk_finite_scalar(x, name)

chk_finite_vector(x, name)

err_finite_scalar(x, name)

err_finite_vector(x, name)
}
\arguments{
\item{x}{A numeric scalar.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a scalar or vector only has finite values
}
\examples{
x <- 0
chk_finite_scalar(x, name = "x")
x <- 1:5
chk_finite_vector(x, name = "x")
}
