% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_first_day_unit_scalar}
\alias{chk_first_day_unit_scalar}
\alias{chk_first_day_unit_vector}
\alias{chk_first_day_unit_consec}
\alias{err_first_day_unit_scalar}
\alias{err_first_day_unit_vector}
\alias{err_first_day_unit_consec}
\title{Check dates are the first day of the specified unit}
\usage{
chk_first_day_unit_scalar(x, name, unit)

chk_first_day_unit_vector(x, name, unit)

chk_first_day_unit_consec(x, name, unit)

err_first_day_unit_scalar(x, name, unit)

err_first_day_unit_vector(x, name, unit)

err_first_day_unit_consec(x, name, unit)
}
\arguments{
\item{x}{A scalar or vector of dates, or values
that can be coerced to dates.}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{unit}{Measurement units for time, eg \code{"month"}.}
}
\description{
If \code{unit} is \code{"months"},
then \code{"2000-03-01"} is the first day of the specified
unit, but \code{"2000-03-02"} is not.
}
\details{
\code{chk_first_day_unit_consec} adds the requirement
that the dates be consecutive.
}
\examples{
x <- "2020-01-01"
chk_first_day_unit_scalar(x, name = "x", unit = "year")
x <- c("2020-01-01", "2020-02-01")
chk_first_day_unit_vector(x, name = "x", unit = "month")
chk_first_day_unit_consec(x, name = "x", unit = "month")
}
\seealso{
\code{\link{chk_is_date_equiv_scalar}}
}
