% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_ge_break_min_age}
\alias{chk_ge_break_min_age}
\alias{err_ge_break_min_age}
\title{Check that implied ages are greater than or equal to 'break_min'}
\usage{
chk_ge_break_min_age(age, break_min, date, dob, unit)

err_ge_break_min_age(age, break_min, date, dob, unit)
}
\arguments{
\item{age}{Integer. Age, typically in years, but can be other unit.}

\item{break_min}{Integer or NULL. If non-NULL, lowest permissible value
for \code{age}.}

\item{date}{Date on which event occurred or measurement made.
Object of class \code{\link[base:Dates]{Date}}.}

\item{dob}{Date of birth. Object of class \code{\link[base:Dates]{Date}}.}

\item{unit}{Measurement units for age, eg \code{"month"}.}
}
\description{
Check that implied ages are greater than or equal to 'break_min'
}
\examples{
age <- c(20L, 23L)
break_min <- 15L
date <- as.Date(c("2020-03-17", "2020-03-18"))
dob <- as.Date(c("2000-02-01", "1997-01-13"))
unit <- "years"
chk_ge_break_min_age(age = age,
                     break_min = break_min,
                     date = date,
                     dob = dob,
                     unit = unit)
}
\seealso{
\code{\link{chk_ge_break_min_date}},
\code{\link{chk_lt_break_max_age}},
\code{\link{chk_lt_break_max_date}}
}
