% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_ge_break_min_date}
\alias{chk_ge_break_min_date}
\alias{err_ge_break_min_date}
\title{Check that dates are greater than or equal to 'break_min'}
\usage{
chk_ge_break_min_date(date, break_min)

err_ge_break_min_date(date, break_min)
}
\arguments{
\item{date}{Date on which event occurred or measurement made.
Object of class \code{\link[base:Dates]{Date}}.}

\item{break_min}{Date or NULL. If non-NULL, lowest permissible value
for \code{date}.}
}
\description{
Check that dates are greater than or equal to 'break_min'
}
\examples{
date <- as.Date(c("2020-03-17", "2020-03-18"))
break_min <- as.Date("2020-01-01")
chk_ge_break_min_date(date = date,
                      break_min = break_min)
}
\seealso{
\code{\link{chk_ge_break_min_age}},
\code{\link{chk_lt_break_max_age}}
\code{\link{chk_lt_break_max_date}}
}
