% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_has_dimnames}
\alias{chk_has_dimnames}
\alias{chk_has_names_dimnames}
\alias{err_has_dimnames}
\alias{err_has_names_dimnames}
\title{Check that an array has dimnames}
\usage{
chk_has_dimnames(x, name)

chk_has_names_dimnames(x, name)

err_has_dimnames(x, name)

err_has_names_dimnames(x, name)
}
\arguments{
\item{x}{An array.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an array has dimnames, and that
those dimnames have names.
}
\examples{
x <- array(1:6,
           dim = 3:2,
           dimnames = list(age = c("0-4", "5-9", "10+"),
                           sex = c("Female", "Male")))
chk_has_dimnames(x, name = "x")
chk_has_names_dimnames(x, name = "x")
}
\seealso{
\code{\link{chk_no_dimnames}},
\code{\link{chk_array_metadata_complete}}
}
