% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_has_non_na}
\alias{chk_has_non_na}
\alias{err_has_non_na}
\title{Check that a vector has at least one
non-NA element}
\usage{
chk_has_non_na(x, name)

err_has_non_na(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector has at least one
non-NA element
}
\examples{
x <- c(0L, NA, NA)
chk_has_non_na(x = x, name = "x")
}
\seealso{
\code{\link{chk_has_nonzero}}
}
