% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_has_nonzero}
\alias{chk_has_nonzero}
\alias{err_has_nonzero}
\title{Check that a vector has at least one
non-zero element}
\usage{
chk_has_nonzero(x, name)

err_has_nonzero(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector has at least one
non-zero element
}
\examples{
x <- c(0L, 1L, 2L, 0L)
chk_has_nonzero(x = x, name = "x")
}
\seealso{
\code{\link{chk_all_0_1}}
}
