% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_increasing}
\alias{chk_increasing}
\alias{chk_strictly_increasing}
\alias{err_increasing}
\alias{err_strictly_increasing}
\title{Check that elements in a vector of
numbers or dates are increasing
or strictly increasing}
\usage{
chk_increasing(x, name)

chk_strictly_increasing(x, name)

err_increasing(x, name)

err_strictly_increasing(x, name)
}
\arguments{
\item{x}{A vector of dates or numbers.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
\code{chk_increasing} checks that a value is equal to
or greater than the one before it;
\code{chk_strictly_increasing} checks that a value is
greater than the one before it.
}
\examples{
x <- c(1:4, 4)
chk_increasing(x, name = "x")
x <- 1:5
chk_strictly_increasing(x, name = "x")
}
