% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R
\name{chk_indices_distinct}
\alias{chk_indices_distinct}
\title{Check that dimension indices all refer
to different dimensions}
\usage{
chk_indices_distinct(indices, names, exclude_zero)
}
\arguments{
\item{indices}{A list of integer vectors
(typically of length 1).}

\item{names}{A character vector with the names
of the indices.}

\item{exclude_zero}{Logical. Whether to
exclude zeros in \code{indices} before
testing for uniqueness.}
}
\description{
Check that dimension indices all refer
to different dimensions
}
\examples{
indices <- list(3L, 1:2, 0L, 6:5)
names <- c("i_time", "i_orig", "i_age", "i_dest")
exclude_zero <- TRUE
chk_indices_distinct(indices = indices,
                     names = names,
                     exclude_zero = exclude_zero)
}
