% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_integer_in_range}
\alias{chk_integer_in_range}
\alias{err_integer_in_range}
\title{Check all elements of an integer vector are
within a given range}
\usage{
chk_integer_in_range(x, min, max, name)

err_integer_in_range(x, min, max, name)
}
\arguments{
\item{x}{An integer vector.}

\item{min}{Minimum permissible value. An integer.}

\item{max}{Maximum permissible value. An integer.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
NAs are permitted
}
\examples{
chk_integer_in_range(x = c(1L, 7L, NA),
                     min = 1L,
                     max = 12L,
                     name = "x")
}
