% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_interval_diff_gt_one}
\alias{chk_interval_diff_gt_one}
\alias{err_interval_diff_gt_one}
\title{Check that difference between upper and lower limits of
interval greater than 1}
\usage{
chk_interval_diff_gt_one(int_low, int_up, is_low_up, labels)

err_interval_diff_gt_one(int_low, int_up, is_low_up, labels)
}
\arguments{
\item{int_low}{An integer vector. The lower
breaks for the intervals.}

\item{int_up}{An integer vector. The upper
breaks for the intervals.}

\item{is_low_up}{Logical. Whether label
has "low-up" format, eg "15-19".}

\item{labels}{A vector of interval labels.}
}
\description{
Check applies to intervals with labels that have
a "low-up" format, ie not single-year age groups,
not open age groups or cohorts, and not missing.
}
\examples{
chk_interval_diff_gt_one(int_low = c(0, 1, 5, 10),
                         int_up = c(1, 5, 10, NA),
                         is_low_up = c(TRUE, TRUE, TRUE, FALSE),
                         labels = c("0", "1-4", "5-9", "10+"))
}
