% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_interval_label_ge_break_min}
\alias{chk_interval_label_ge_break_min}
\alias{chk_interval_label_le_break_max}
\alias{err_interval_label_ge_break_min}
\alias{err_interval_label_le_break_max}
\title{Check that all intervals fall within limits
imposed by 'break_min' or 'break_max'}
\usage{
chk_interval_label_ge_break_min(labels, int_low, break_min)

chk_interval_label_le_break_max(labels, int_up, break_max)

err_interval_label_ge_break_min(labels, int_low, break_min)

err_interval_label_le_break_max(labels, int_up, break_max)
}
\arguments{
\item{labels}{A vector of interval labels.}

\item{int_low}{An integer vector. The lower
breaks for the intervals.}

\item{break_min}{Integer. Lower limit for all intervals,
expect for left open interval if present.}

\item{int_up}{An integer vector. The upper
breaks for the intervals.}

\item{break_max}{Integer. Upper limit for all intervals,
except for right open interval if preseent.}
}
\description{
All age groups are assumed to be closed.
}
\examples{
chk_interval_label_ge_break_min(labels = c("15-19", "20", "100-104"),
                                int_low = c(15, 20, 100, NA),
                                break_min = 15)
chk_interval_label_le_break_max(labels = c("20-24", "45-49", NA),
                                int_up = c(25, 50, NA),
                                break_max = 50)
}
