% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_character}
\alias{chk_is_character}
\alias{err_is_character}
\title{Check that a value has type "character"}
\usage{
chk_is_character(x, name)

err_is_character(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a value has type "character"
}
\examples{
x <- c("a", "b", "c")
chk_is_character(x, name = "x")
}
