% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_is_class}
\alias{chk_is_class}
\alias{chk_is_class_obj}
\alias{chk_is_class_list}
\alias{err_is_class_obj}
\alias{err_is_class_list}
\title{Check that object or elements of list inherit
from specified class}
\usage{
chk_is_class_obj(x, name, class)

chk_is_class_list(x, name, class)

err_is_class_obj(x, name, class)

err_is_class_list(x, name, class)
}
\arguments{
\item{x}{An object, or a list of objects}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{class}{Name of a class.}
}
\description{
Check that object or elements of list inherit
from specified class
}
\examples{
x <- "a"
class <- "character"
chk_is_class_obj(x, name = "x", class = class)
x <- list(c("a", "b"), "c")
chk_is_class_list(x, name = "x", class = class)
}
