% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_is_class_scalar}
\alias{chk_is_class_scalar}
\alias{chk_is_class_vector}
\alias{err_is_class_scalar}
\alias{err_is_class_vector}
\title{Chech that scalar or vector inherits from specified class}
\usage{
chk_is_class_scalar(x, name, class)

chk_is_class_vector(x, name, class)

err_is_class_scalar(x, name, class)

err_is_class_vector(x, name, class)
}
\arguments{
\item{x}{A scalar or vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{class}{Name of a class.}
}
\description{
Chech that scalar or vector inherits from specified class
}
\examples{
x <- "a"
class <- "character"
chk_is_class_scalar(x, name = "x", class = class)
x <- c("a", "b")
chk_is_class_vector(x, name = "x", class = class)
}
