% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_date}
\alias{chk_is_date}
\alias{err_is_date}
\title{Check that a value inherits from class "Date"}
\usage{
chk_is_date(x, name)

err_is_date(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a value inherits from class "Date"
}
\examples{
x <- as.Date(c("2020-01-01", "2021-01-01"))
chk_is_date(x, name = "x")
}
