% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_date_equiv_scalar}
\alias{chk_is_date_equiv_scalar}
\alias{chk_is_date_equiv_vector}
\alias{err_is_date_equiv_scalar}
\alias{err_is_date_equiv_vector}
\title{Check that a scalar or vector is dates or
can be coerced to dates}
\usage{
chk_is_date_equiv_scalar(x, name)

chk_is_date_equiv_vector(x, name)

err_is_date_equiv_scalar(x, name)

err_is_date_equiv_vector(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a scalar or vector is dates or
can be coerced to dates
}
\examples{
x <- as.Date("2001-12-31")
chk_is_date_equiv_scalar(x, name = "x")
x <- as.Date(c("2001-12-31", "2002-01-01"))
chk_is_date_equiv_vector(x, name = "x")
}
\seealso{
\code{\link{chk_is_date}} for a stricter test,
and \code{\link{err_tdy_date_scalar}} and
\code{\link{err_tdy_date_vector}} to coerce
where necessary.
}
