% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_date_or_numeric}
\alias{chk_is_date_or_numeric}
\alias{err_is_date_or_numeric}
\title{Check that a vector is dates or numbers}
\usage{
chk_is_date_or_numeric(x, name)

err_is_date_or_numeric(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
The vector can have length 1,
ie be a scalar.
}
\examples{
x <- as.Date("2001-03-01")
chk_is_date_or_numeric(x, name = "x")
x <- 5:1
chk_is_date_or_numeric(x, name = "x")
}
