% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_factor}
\alias{chk_is_factor}
\alias{err_is_factor}
\title{Check that a value has type "factor"}
\usage{
chk_is_factor(x, name)

err_is_factor(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a value has type "factor"
}
\examples{
x <- factor(c("a", "b", "c"))
chk_is_factor(x, name = "x")
}
