% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_integer}
\alias{chk_is_integer}
\alias{chk_is_integer_consec}
\alias{err_is_integer}
\alias{err_is_integer_consec}
\title{Check that a scalar or vector has type integer}
\usage{
chk_is_integer(x, name)

chk_is_integer_consec(x, name)

err_is_integer(x, name)

err_is_integer_consec(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checking type, e.g. fails for 1.0.
\code{chk_is_integer_conse} adds the condition
that the integers be consecutive.
}
\examples{
x <- 1:3
chk_is_integer(x, name = "x")
chk_is_integer_consec(x, name = "x")
}
\seealso{
\code{\link{chk_is_integer_equiv_scalar}} for looser
integer test. \code{\link{chk_is_numeric}} to test for
numeric type.
}
