% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_integer_equiv_scalar}
\alias{chk_is_integer_equiv_scalar}
\alias{chk_is_integer_equiv_vector}
\alias{err_is_integer_equiv_scalar}
\alias{err_is_integer_equiv_vector}
\title{Check that a scalar or vector is integer or can be coerced to integer}
\usage{
chk_is_integer_equiv_scalar(x, name)

chk_is_integer_equiv_vector(x, name)

err_is_integer_equiv_scalar(x, name)

err_is_integer_equiv_vector(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a scalar or vector is integer or can be coerced to integer
}
\examples{
x <- 1.0
chk_is_integer_equiv_scalar(x, name = "x")
x <- 1:4
chk_is_integer_equiv_vector(x, name = "x")
}
\seealso{
\code{\link{chk_is_integer}} for a stricter test,
and \code{\link{err_tdy_integer_scalar}} and
\code{\link{err_tdy_integer_vector}} to coerce
to integer where necessary.
}
