% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_list}
\alias{chk_is_list}
\alias{err_is_list}
\title{Check that a vector has class "list"}
\usage{
chk_is_list(x, name)

err_is_list(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector has class "list"
}
\examples{
x <- list(1L, "a", NA)
chk_is_list(x, name = "x")
}
