% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_is_logical_flag}
\alias{chk_is_logical_flag}
\alias{err_is_logical_flag}
\title{Check that 'x' is a non-NA logical vector of length 1}
\usage{
chk_is_logical_flag(x, name)

err_is_logical_flag(x, name)
}
\arguments{
\item{x}{A scalar.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that 'x' is a non-NA logical vector of length 1
}
\examples{
x <- TRUE
chk_is_logical_flag(x, name = "x")
}
