% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_is_na}
\alias{chk_is_na}
\alias{chk_is_na_scalar}
\alias{chk_is_na_vector}
\alias{err_is_na_scalar}
\alias{err_is_na_vector}
\title{Check that a scalar is NA, or a vector
is all NA}
\usage{
chk_is_na_scalar(x, name)

chk_is_na_vector(x, name)

err_is_na_scalar(x, name)

err_is_na_vector(x, name)
}
\arguments{
\item{x}{A scalar or a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a scalar is NA, or a vector
is all NA
}
\examples{
x <- NA
chk_is_na_scalar(x, name = "x")
x <- c(NA_integer_, NA_integer_)
chk_is_na_vector(x, name = "x")
}
\seealso{
\code{\link{chk_not_na}}
}
