% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_is_string}
\alias{chk_is_string}
\alias{err_is_string}
\title{Check that an object is a non-NA character
vector of length 1}
\usage{
chk_is_string(x, name)

err_is_string(x, name)
}
\arguments{
\item{x}{An object.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an object is a non-NA character
vector of length 1
}
\examples{
x <- "hello"
chk_is_string(x, name = "x")
}
