% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_items_date}
\alias{chk_items_date}
\alias{err_items_date}
\title{Check that all items in a list are Date vectors}
\usage{
chk_items_date(x, name)

err_items_date(x, name)
}
\arguments{
\item{x}{A list.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that all items in a list are Date vectors
}
\examples{
x <- list(as.Date(character()),
          as.Date(c("2000-01-01", "2001-02-01")))
chk_items_date(x = x,
               name = "x")
}
