% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_items_increasing}
\alias{chk_items_increasing}
\alias{err_items_increasing}
\title{Check that elements within each item
are increasing or strictly increasing}
\usage{
chk_items_increasing(x, strict, name)

err_items_increasing(x, strict, name)
}
\arguments{
\item{x}{A list.}

\item{strict}{Logical. Whether elements should be
strictly increasing}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
NAs are ignored.
}
\examples{
x <- list(c(0L, c(1L, 1L, 2L), 3:1, c(NA, 5, 10)))
chk_items_increasing(x = x,
                     strict = FALSE,
                     name = "x")
x <- list(c(0L, c(1L, NA, 2L), 3:1, c(NA, 5, 10)))
chk_items_increasing(x = x,
                     strict = TRUE,
                     name = "x")
}
