% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_items_integer}
\alias{chk_items_integer}
\alias{err_items_integer}
\title{Check that all items in a list are integer vectors}
\usage{
chk_items_integer(x, name)

err_items_integer(x, name)
}
\arguments{
\item{x}{A list.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that all items in a list are integer vectors
}
\examples{
x <- list(integer(), 1:3, 0L)
chk_items_integer(x = x,
                  name = "x")
}
