% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_items_length_k}
\alias{chk_items_length_k}
\alias{err_items_length_k}
\title{Check that all items in a list have length k}
\usage{
chk_items_length_k(x, k, name)

err_items_length_k(x, k, name)
}
\arguments{
\item{x}{A list.}

\item{k}{The required length of each item.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that all items in a list have length k
}
\examples{
x <- list(3:1, 1:3, c("a", "b", "c"))
chk_items_length_k(x = x,
                   k = 3L,
                   name = "x")
}
