% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_items_no_na}
\alias{chk_items_no_na}
\alias{err_items_no_na}
\title{Check that not items have NAs, except where
specified in 'except' argument}
\usage{
chk_items_no_na(x, except, name)

err_items_no_na(x, except, name)
}
\arguments{
\item{x}{A list.}

\item{except}{A list of integer vectors, each of which is length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that not items have NAs, except where
specified in 'except' argument
}
\examples{
x <- list(c(0L, NA, 1L), 3:1, c(NA, "b", "c"), 1L)
chk_items_no_na(x = x,
                except = list(c(1L, 2L), c(3L, 1L)),
                name = "x")
}
