% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_items_one_greater}
\alias{chk_items_one_greater}
\alias{err_items_one_greater}
\title{Check that, in items of list, second element
one greater than first}
\usage{
chk_items_one_greater(x, name)

err_items_one_greater(x, name)
}
\arguments{
\item{x}{A list.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
NAs are ignored.
}
\examples{
x <- list(c(0L, 1L), c(1L, 2L), c(NA, 100L), c(NA, 5L))
chk_items_one_greater(x = x,
                      name = "x")
}
