% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-label_values.R
\name{chk_label_values_categories}
\alias{chk_label_values_categories}
\title{Check values for Categories labels}
\usage{
chk_label_values_categories(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is character
  \item at most one NA
  \item no blanks (zero-length strings)
  \item no duplicates
}
}
\examples{
x <- c("Thailand", "Laos", "Cambodia")
chk_label_values_categories(x = x, name = "x")
}
