% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-label_values.R
\name{chk_label_values_direction}
\alias{chk_label_values_direction}
\title{Check values for Direction labels}
\usage{
chk_label_values_direction(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is character
  \item has values \code{c("In", "Out")}, not necessarily in that order
}
}
\examples{
x <- c("In", "Out")
chk_label_values_direction(x = x, name = "x")
}
