% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-label_values.R
\name{chk_label_values_integers}
\alias{chk_label_values_integers}
\title{Check values for Integers labels}
\usage{
chk_label_values_integers(x, name)
}
\arguments{
\item{x}{An integer vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is integer
  \item no duplicates
  \item no NAs
}
}
\examples{
x <- c(0L, 5L, 22L, 23L)
chk_label_values_integers(x = x, name = "x")
}
