% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-label_values.R
\name{chk_label_values_quantiles}
\alias{chk_label_values_quantiles}
\title{Check values for Quantiles labels}
\usage{
chk_label_values_quantiles(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is character
  \item no NAs
  \item all elements have format "<number>%" where 0 <= number <= 100
}
}
\examples{
x <- c("2.5\%", "50\%", "97.5\%")
chk_label_values_quantiles(x = x, name = "x")
}
