% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-label_values.R
\name{chk_label_values_quantities}
\alias{chk_label_values_quantities}
\title{Check values for Quantities labels}
\usage{
chk_label_values_quantities(x, name)
}
\arguments{
\item{x}{A list of integer vectors, 
each of which has length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is list of integer vectors of length 2
  \item no NAs, except element 1 of first item, and element 2 of last item
  \item second element greater than or equal to first
  \item pairs do not overlap
}
}
\examples{
x <- list(c(100L, NA), c(0L, 4L), c(NA, -1L), c(5L, 49L))
chk_label_values_quantities(x = x, name = "x")
}
