% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_le}
\alias{chk_le}
\alias{chk_le_scalar}
\alias{chk_le_vector}
\alias{err_le_scalar}
\alias{err_le_vector}
\title{Check that 'x1' is less than or equal to 'x2'}
\usage{
chk_le_scalar(x1, x2, name1, name2)

chk_le_vector(x1, x2, name1, name2)

err_le_scalar(x1, x2, name1, name2)

err_le_vector(x1, x2, name1, name2)
}
\arguments{
\item{x1}{A scalar or vector.}

\item{x2}{A scalar or vector the same length as \code{x1},
or length 1.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
Check that 'x1' is less than or equal to 'x2'
}
\examples{
x1 <- 2.2
x2 <- 3.1
chk_le_scalar(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- c(3.3, 4.2, 5.7)
x2 <- c(3.3, 4.6, 6.2)
chk_le_vector(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- c(3.3, 4.2, 5.7)
x2 <- 3.3
chk_le_vector(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
}
\seealso{
\code{\link{chk_lt}}, \code{\link{chk_ge}},
\code{\link{chk_gt}}
}
