% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_length_1}
\alias{chk_length_1}
\alias{err_length_1}
\title{Check that a vector has length 1,
ie is a scalar}
\usage{
chk_length_1(x, name)

err_length_1(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector has length 1,
ie is a scalar
}
\examples{
x <- "a"
chk_length_1(x, name = "x")
}
