% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_length_equals}
\alias{chk_length_equals}
\alias{err_length_equals}
\title{Check that 'x1' has length specified by 'x2'}
\usage{
chk_length_equals(x1, x2, name1, name2)

err_length_equals(x1, x2, name1, name2)
}
\arguments{
\item{x1}{An object.}

\item{x2}{A positive integer.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
Check that 'x1' has length specified by 'x2'
}
\examples{
chk_length_equals(x1 = c(1L, 3L, 2L),
                  x2 = 3L,
                  name1 = "x1",
                  name2 = "x2")
}
\seealso{
\code{\link{chk_length_same_or_1}},
\code{\link{chk_length_same}}
}
