% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_length_same}
\alias{chk_length_same}
\alias{err_length_same}
\title{Check that 'x1' and 'x2' have same length}
\usage{
chk_length_same(x1, x2, name1, name2)

err_length_same(x1, x2, name1, name2)
}
\arguments{
\item{x1}{An object.}

\item{x2}{An object.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
Check that 'x1' and 'x2' have same length
}
\examples{
x1 <- 1:5
x2 <- 2:6
chk_length_same(x1 = x1, x2 = x2,
                name1 = "x1", name2 = "x2")
}
\seealso{
\code{\link{chk_length_equals}},
\code{\link{chk_length_same_or_1}}
}
