% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_lengths_elements_equal_vec}
\alias{chk_lengths_elements_equal_vec}
\alias{err_lengths_elements_equal_vec}
\title{Check that 'x2' gives the lengths of the elements of 'x1'}
\usage{
chk_lengths_elements_equal_vec(x1, x2, name1, name2)

err_lengths_elements_equal_vec(x1, x2, name1, name2)
}
\arguments{
\item{x1}{A list.}

\item{x2}{An integer vector, the same length as \code{x1}.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
Check that 'x2' gives the lengths of the elements of 'x1'
}
\examples{
x1 <- list(1:4, 1:2, 1:3)
x2 <- c(4L, 2L, 3L)
chk_lengths_elements_equal_vec(x1 = x1, x2 = x2,
                               name1 = "x1", name2 = "x2")
}
