% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_lt_break_max_age}
\alias{chk_lt_break_max_age}
\alias{err_lt_break_max_age}
\title{Check that implied ages are less than 'break_max'}
\usage{
chk_lt_break_max_age(age, break_max, date, dob, unit)

err_lt_break_max_age(age, break_max, date, dob, unit)
}
\arguments{
\item{age}{Integer. Age, typically in years, but can be other unit.}

\item{break_max}{Integer or NULL. If non-NULL, upper limit
for \code{age}.}

\item{date}{Date on which event occurred or measurement made.
Object of class \code{\link[base:Dates]{Date}}.}

\item{dob}{Date of birth. Object of class \code{\link[base:Dates]{Date}}.}

\item{unit}{Measurement units for age, eg \code{"month"}.}
}
\description{
Check that implied ages are less than 'break_max'
}
\examples{
age <- c(80L, 83L)
break_max <- 85L
date <- as.Date(c("2020-03-17", "2020-03-18"))
dob <- as.Date(c("1940-02-01", "1937-01-13"))
unit <- "years"
chk_lt_break_max_age(age = age,
                     break_max = break_max,
                     date = date,
                     dob = dob,
                     unit = unit)
}
\seealso{
\code{\link{chk_ge_break_min_age}},
\code{\link{chk_ge_break_min_date}},
\code{\link{chk_lt_break_max_date}}
}
