% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_lt_break_max_date}
\alias{chk_lt_break_max_date}
\alias{err_lt_break_max_date}
\title{Check that dates are less than 'break_max'}
\usage{
chk_lt_break_max_date(date, break_max)

err_lt_break_max_date(date, break_max)
}
\arguments{
\item{date}{Date on which event occurred or measurement made.
Object of class \code{\link[base:Dates]{Date}}.}

\item{break_min}{Date or NULL. If non-NULL, lowest permissible value
for \code{date}.}
}
\description{
Check that dates are less than 'break_max'
}
\examples{
date <- as.Date(c("2020-03-17", "2020-03-18"))
break_max <- as.Date("2020-06-01")
chk_lt_break_max_date(date = date, break_max = break_max)
}
\seealso{
\code{\link{chk_ge_break_min_age}},
\code{\link{chk_ge_break_min_date}},
\code{\link{chk_lt_break_max_age}}
}
