% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_map_dim}
\alias{chk_map_dim}
\alias{err_map_dim}
\title{Check 'map_dim'}
\usage{
chk_map_dim(map_dim, dim_self, dim_oth)

err_map_dim(map_dim, dim_self, dim_oth)
}
\arguments{
\item{map_dim}{A vector of non-negative integers,
unique apart from any zeros, the same length as
\code{dim_self}.}

\item{dim_self}{The dimensions of \code{self}.}

\item{dim_oth}{The dimensions of \code{oth}.}
}
\description{
Check that a \code{\link[=err_tdy_map_dim]{map_dim}}
argument is valid.
}
\examples{
map_dim <- c(3, 1, 0)
dim_self <- c(4L, 6L, 2L)
dim_oth <- c(6L, 3L, 4L, 3L)
chk_map_dim(map_dim = map_dim,
            dim_self = dim_self,
            dim_oth = dim_oth)
}
