% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_map_pos}
\alias{chk_map_pos}
\alias{err_map_pos}
\title{Check 'map_dim'}
\usage{
chk_map_pos(x, name)

err_map_pos(x, name)
}
\arguments{
\item{x}{A list of integer vectors.}

\item{name}{The name for \code{x} to be used
in error messages.}
}
\description{
Validity tests for a \code{\link[=err_tdy_map_pos]{map_pos}}
that only require the object itself.
}
\examples{
x <- list(c(2L, 0L), 1:3)
chk_map_pos(x = x,
            name = "x")
}
\seealso{
\code{\link{chk_map_pos}}
}
