% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-member.R, R/err-member.R
\name{chk_member_comp_type}
\alias{chk_member_comp_type}
\alias{err_member_comp_type}
\title{Check that 'x' is a valid type of account
component}
\usage{
chk_member_comp_type(x, name)

err_member_comp_type(x, name)
}
\arguments{
\item{x}{A string.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that 'x' is a valid type of account
component
}
\examples{
x <- "orig-dest"
chk_member_comp_type(x, name = "x")
}
