% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-member.R, R/err-member.R
\name{chk_member_dimtype}
\alias{chk_member_dimtype}
\alias{err_member_dimtype}
\title{Check that 'x' consists of valid dimtypes}
\usage{
chk_member_dimtype(x, name)

err_member_dimtype(x, name)
}
\arguments{
\item{x}{A character vector}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that 'x' consists of valid dimtypes
}
\examples{
x <- c("age", "attribute")
chk_member_dimtype(x, name = "x")
}
