% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-member.R, R/err-member.R
\name{chk_member_unit}
\alias{chk_member_unit}
\alias{err_member_unit}
\title{Check that 'x' is a valid time unit}
\usage{
chk_member_unit(x, name)

err_member_unit(x, name)
}
\arguments{
\item{x}{A string - ie a character vector of length 1.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that 'x' is a valid time unit
}
\examples{
x <- "year"
chk_member_unit(x = x, name = "x")
}
