% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_multiple_of_n}
\alias{chk_multiple_of_n}
\alias{err_multiple_of_n}
\title{Check that 'x1' is a multiple of 'n'}
\usage{
chk_multiple_of_n(x, name, n)

err_multiple_of_n(x, name, n)
}
\arguments{
\item{x}{A scalar.}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{n}{A scalar.}
}
\description{
\code{chk_multiple_of_n} differs from
\code{\link{chk_multiple_of}} only in the
error message. \code{chk_multiple_of_n} refers
to the value of \code{n}, rather than to the
variable \code{"n"}.
}
\examples{
x <- 10L
n <- 2L
chk_multiple_of_n(x = x, n = n, name = "x")
}
