% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_names_complete}
\alias{chk_names_complete}
\alias{err_names_complete}
\title{Check that an object has a complete set of names}
\usage{
chk_names_complete(x, name)

err_names_complete(x, name)
}
\arguments{
\item{x}{An object, typically a vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an object has a complete set of names
}
\examples{
x <- c(a = 1, b = 1.2)
chk_names_complete(x = x, name = "x")#' 
}
\seealso{
\code{\link{chk_names_dimnames_complete}},
\code{\link{chk_has_dimnames}},
\code{\link{chk_array_metadata_complete}}
}
