% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_names_dimnames_complete}
\alias{chk_names_dimnames_complete}
\alias{err_names_dimnames_complete}
\title{Check that names for an
array are complete}
\usage{
chk_names_dimnames_complete(x, name)

err_names_dimnames_complete(x, name)
}
\arguments{
\item{x}{An array.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that names for an
array are complete
}
\examples{
x <- array(1:6,
           dim = 2:3,
           dimnames = list(sex = c("Female", "Male"),
                           region = c("A", "B", "C")))
chk_names_dimnames_complete(x = x, name = "x")
}
\seealso{
\code{\link{chk_names_complete}},
\code{\link{chk_has_dimnames}},
\code{\link{chk_array_metadata_complete}}
}
