% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_names_pairs_complete}
\alias{chk_names_pairs_complete}
\alias{err_names_pairs_complete}
\title{Check that dimensions that should have pairs do
in fact have them}
\usage{
chk_names_pairs_complete(names)

err_names_pairs_complete(names)
}
\arguments{
\item{names}{A character vector of dimension names.}
}
\description{
Function \code{\link{chk_names_pairs_suffix}}
should be run before this one.
}
\examples{
names <- c("reg_dest", "reg_orig")
chk_names_pairs_complete(names)
}
