% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_names_pairs_suffix}
\alias{chk_names_pairs_suffix}
\alias{err_names_pairs_suffix}
\title{Check that dimension names have pair suffixes if and only if
they have the appropriate dimtypes}
\usage{
chk_names_pairs_suffix(dimtypes, names)

err_names_pairs_suffix(dimtypes, names)
}
\arguments{
\item{dimtypes}{A character vector of dimtypes.}

\item{names}{A character vector of names.}
}
\description{
Check that dimension names have pair suffixes if and only if
they have the appropriate dimtypes
}
\examples{
dimtypes <- c("origin", "destination")
names <- c("reg_orig", "reg_dest")
chk_names_pairs_suffix(dimtypes = dimtypes, names = names)
}
\seealso{
Function \code{\link{chk_names_pairs_complete}} should only
be called if \code{chk_names_pairs_suffix} has been
called first
}
