% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_no_dimnames}
\alias{chk_no_dimnames}
\alias{err_no_dimnames}
\title{Check that an array does not have dimnames}
\usage{
chk_no_dimnames(x, name)

err_no_dimnames(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an array does not have dimnames
}
\examples{
x <- array(1:6, dim = 3:2)
chk_no_dimnames(x, name = "x")
}
\seealso{
\code{\link{chk_has_dimnames}} to test
that an array does have dimnames. \code{\link{chk_no_names}}
to test that an object does not have names.
}
