% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_no_duplicates}
\alias{chk_no_duplicates}
\alias{err_no_duplicates}
\title{Check that a vector has no duplicates}
\usage{
chk_no_duplicates(x, name)

err_no_duplicates(x, name)
}
\arguments{
\item{x}{A character scalar or vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector has no duplicates
}
\examples{
x <- c(0L, 5L, 10L, NA)
chk_no_duplicates(x, name = "x")
}
