% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_no_names}
\alias{chk_no_names}
\alias{err_no_names}
\title{Check that an object does not have names}
\usage{
chk_no_names(x, name)

err_no_names(x, name)
}
\arguments{
\item{x}{An object.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an object does not have names
}
\examples{
x <- 1:3
chk_no_names(x, name = "x")
}
\seealso{
\code{\link{chk_no_dimnames}} to test
that an array does not have dimnames.
}
