% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_no_open_age}
\alias{chk_no_open_age}
\alias{err_no_open_age}
\title{Check that labels do not have any open age groups,
assuming that 'open_last' is FALSE}
\usage{
chk_no_open_age(x)

err_no_open_age(x)
}
\arguments{
\item{x}{A character vector with age labels.}
}
\description{
Check that labels do not have any open age groups,
assuming that 'open_last' is FALSE
}
\examples{
chk_no_open_age(c("0-4", "5", "100"))

}
