% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_no_open_cohort}
\alias{chk_no_open_cohort}
\alias{err_no_open_cohort}
\title{Check that labels do not have any open cohorts,
assuming that 'open_first' is FALSE}
\usage{
chk_no_open_cohort(x)

err_no_open_cohort(x)
}
\arguments{
\item{x}{A character vector with cohort labels.}
}
\description{
Check that labels do not have any open cohorts,
assuming that 'open_first' is FALSE
}
\examples{
chk_no_open_cohort(c("2000-2005", "2005-2010", "2010-2015"))

}
