% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_no_open_first}
\alias{chk_no_open_first}
\alias{err_no_open_first}
\title{Check that calendar labels not open on left}
\usage{
chk_no_open_first(x, name, unit)

err_no_open_first(x, name, unit)
}
\arguments{
\item{x}{A character vector of calender interval labels.}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{unit}{The time unit: \code{"year"}, \code{"quarter"}
or \code{"month"}.}
}
\description{
Check that calendar labels not open on left
}
\examples{
chk_no_open_first(x = c("2000 Jan", "2005 Aug", "2006 Nov"),
                     name = "x",
                     unit = "month")
chk_no_open_first(c("2000 Q1", "2005 Q3", "2006 Q2"),
                     name = "x",
                     unit = "quarter")
chk_no_open_first(c("2000", "2005", "2006"),
                     name = "x",
                     unit = "year")
}
