% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_non_negative}
\alias{chk_non_negative}
\alias{chk_non_negative_scalar}
\alias{chk_non_negative_vector}
\alias{err_non_negative_scalar}
\alias{err_non_negative_vector}
\title{Check that elements of a scalar or vector are non-negative}
\usage{
chk_non_negative_scalar(x, name)

chk_non_negative_vector(x, name)

err_non_negative_scalar(x, name)

err_non_negative_vector(x, name)
}
\arguments{
\item{x}{A scalar or vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
NAs are ignored.
}
\examples{
x <- 0.1
chk_non_negative_scalar(x, name = "x")
x <- NA_integer_
chk_non_negative_scalar(x, name = "x")
x <- c(0.1, 0, NA)
chk_non_negative_vector(x, name = "x")
}
\seealso{
\code{\link{chk_positive_scalar}}
}
