% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_nonzero_unique}
\alias{chk_nonzero_unique}
\alias{err_nonzero_unique}
\title{Check that non-zero elements in an integer
vector are unique}
\usage{
chk_nonzero_unique(x, name)

err_nonzero_unique(x, name)
}
\arguments{
\item{x}{An object.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that non-zero elements in an integer
vector are unique
}
\examples{
x <- c(0L, 2L, 0L, 1L, 0L)
chk_nonzero_unique(x, name = "x")

}
