% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_not_blank}
\alias{chk_not_blank}
\alias{chk_not_blank_scalar}
\alias{chk_not_blank_vector}
\alias{err_not_blank_scalar}
\alias{err_not_blank_vector}
\title{Check that a character scalar or vector is not blank}
\usage{
chk_not_blank_scalar(x, name)

chk_not_blank_vector(x, name)

err_not_blank_scalar(x, name)

err_not_blank_vector(x, name)
}
\arguments{
\item{x}{A character scalar or vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
An element is blank if it equals "".
}
\examples{
x <- "A"
chk_not_blank_scalar(x, name = "x")
x <- c("A", "B", "C")
chk_not_blank_vector(x, name = "x")
}
