% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_not_equal_integer_scalar}
\alias{chk_not_equal_integer_scalar}
\alias{err_not_equal_integer_scalar}
\title{Check that integers 'x1' and 'x2' are not equal}
\usage{
chk_not_equal_integer_scalar(x1, x2, name1, name2)

err_not_equal_integer_scalar(x1, x2, name1, name2)
}
\arguments{
\item{x1}{An integer scalar.}

\item{x2}{An integer scalar.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
Check that integers 'x1' and 'x2' are not equal
}
\examples{
x1 <- 2L
x2 <- 3L
chk_not_equal_integer_scalar(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
}
\seealso{
\code{\link{chk_le}},
\code{\link{chk_lt}}, \code{\link{chk_ge}},
\code{\link{chk_gt}}
}
