% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_not_na_dataframe}
\alias{chk_not_na_dataframe}
\alias{chk_not_na_list}
\alias{chk_not_na_scalar}
\alias{chk_not_na_vector}
\alias{err_not_na_dataframe}
\alias{err_not_na_list}
\alias{err_not_na_scalar}
\alias{err_not_na_vector}
\title{Check that an object does not have NAs}
\usage{
chk_not_na_dataframe(x, name)

chk_not_na_list(x, name)

chk_not_na_scalar(x, name)

chk_not_na_vector(x, name)

err_not_na_dataframe(x, name)

err_not_na_list(x, name)

err_not_na_scalar(x, name)

err_not_na_vector(x, name)
}
\arguments{
\item{x}{A data frame, list,
scalar, or vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an object does not have NAs
}
\examples{
x <- data.frame(a = 1:2, b = 3:4)
chk_not_na_dataframe(x, name = "x")
x <- list(a = 1:2, b = 3:4)
chk_not_na_list(x, name = "x")
x <- 1.3
chk_not_na_scalar(x, name = "x")
x <- 4:1
chk_not_na_vector(x, name = "x")
}
