% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_null}
\alias{chk_null}
\alias{chk_null_if_null}
\alias{chk_null_onlyif_null}
\alias{chk_null_ifonlyif_null}
\alias{err_null_if_null}
\alias{err_null_onlyif_null}
\alias{err_null_ifonlyif_null}
\title{Check NULL status of 'x1' given NULL status of 'x2'}
\usage{
chk_null_if_null(x1, x2, name1, name2)

chk_null_onlyif_null(x1, x2, name1, name2)

chk_null_ifonlyif_null(x1, x2, name1, name2)

err_null_if_null(x1, x2, name1, name2)

err_null_onlyif_null(x1, x2, name1, name2)

err_null_ifonlyif_null(x1, x2, name1, name2)
}
\arguments{
\item{x1}{An argument that could be \code{NULL}.}

\item{x2}{An argument that could be \code{NULL}.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
\code{chk_null_if_null} checks that \code{x1} is \code{NULL},
given that \code{x2} is \code{NULL}.
}
\details{
\code{chk_null_onlyif_null} checks that \code{x1} is only
\code{NULL} when \code{x2} is \code{NULL}.

\code{chk_null_ifonlyif_null} checks that \code{x1} and
\code{x2} are both \code{NULL} or are both non-\code{NULL}.
}
\examples{
x1 <- NULL
x2 <- 2
chk_null_if_null(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- NULL
x2 <- NULL
chk_null_if_null(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 1
x2 <- NULL
chk_null_onlyif_null(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- NULL
x2 <- NULL
chk_null_onlyif_null(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 1
x2 <- 2
chk_null_ifonlyif_null(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- NULL
x2 <- NULL
chk_null_ifonlyif_null(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
}
