% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_omitted}
\alias{chk_omitted}
\alias{chk_not_omitted}
\alias{err_omitted}
\alias{err_not_omitted}
\title{Check whether dimension(s) identified by an index are omitted,
or not omitted, according to 'map_dim'}
\usage{
chk_omitted(index, map_dim, name_index, name_dim)

chk_not_omitted(index, map_dim, name_index, name_dim)

err_omitted(index, map_dim, name_index, name_dim)

err_not_omitted(index, map_dim, name_index, name_dim)
}
\arguments{
\item{index}{An integer vector, identifying dimensions of array \code{self}}

\item{map_dim}{Integer vector mapping dimensions of array \code{self} to
dimensions of array \code{oth}.}

\item{name_index}{Name of \code{index}.}

\item{name_dim}{Name of the dimension that should or should not be omitted.}
}
\description{
\code{chk_omitted} checks that a dimension \emph{is} omitted,
and \code{chk_not_omitted} checks that a dimension \emph{is not}
omitted.
}
\examples{
chk_omitted(index = 2L,
              map_dim = c(1L, 0L, 2L),
              name_index = "indices_orig_self",
              name_dim = "origin")
chk_not_omitted(index = 3L,
                  map_dim = c(1L, 0L, 2L),
                  name_index = "indices_dest_self",
                  name_dim = "destination")
}
\seealso{
\code{\link{chk_map_dim}}
}
