% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_open_left_le_break_min}
\alias{chk_open_left_le_break_min}
\alias{chk_open_right_ge_break_max}
\alias{err_open_left_le_break_min}
\alias{err_open_right_ge_break_max}
\title{Check that open intervals do not overlap with 'break_min' or 'break_max'}
\usage{
chk_open_left_le_break_min(labels, int_up, is_open, break_min)

chk_open_right_ge_break_max(labels, int_low, is_open, break_max)

err_open_left_le_break_min(labels, int_up, is_open, break_min)

err_open_right_ge_break_max(labels, int_low, is_open, break_max)
}
\arguments{
\item{labels}{A character vector with age labels.}

\item{int_up}{Upper limits of intervals.}

\item{is_open}{Logical vector - whether each interval is open}

\item{break_min}{Lowest break.}

\item{int_low}{Lower limits of intervals.}

\item{break_max}{Highest break.}
}
\description{
\code{chk_open_left_le_break_min} assumes that intervals are open
on the left, and checks that they do not exceed \code{break_min}.
}
\details{
\code{chk_open_right_ge_break_max} assumes that intervals open
on the right, and checks that they do not exceed \code{break_max}.
}
\examples{
chk_open_left_le_break_min(labels = c("2005-2010", "<2000", NA),
                           int_up = c(2010, 2000, NA),
                           is_open = c(FALSE, TRUE, FALSE),
                           break_min = 2000)
chk_open_right_ge_break_max(labels = c("2005-2010", "2010+", NA),
                            int_low = c(2005, 2010, NA),
                            is_open = c(FALSE, TRUE, FALSE),
                            break_max = 2010)
}
