% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_positive}
\alias{chk_positive}
\alias{chk_positive_scalar}
\alias{chk_positive_vector}
\alias{err_positive_scalar}
\alias{err_positive_vector}
\title{Check that elements of a scalar or vector are positive}
\usage{
chk_positive_scalar(x, name)

chk_positive_vector(x, name)

err_positive_scalar(x, name)

err_positive_vector(x, name)
}
\arguments{
\item{x}{A scalar or vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
NAs are ignored.
}
\examples{
x <- 0.1
chk_positive_scalar(x, name = "x")
x <- NA_integer_
chk_positive_scalar(x, name = "x")
x <- c(0.1, 0.0001, 100, NA)
chk_positive_vector(x, name = "x")
}
\seealso{
\code{\link{chk_non_negative_scalar}}
}
