% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R
\name{chk_positive_dim}
\alias{chk_positive_dim}
\title{Check a 'dim' argument for an array with
positive length}
\usage{
chk_positive_dim(x, name)
}
\arguments{
\item{x}{An integer vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that \code{x} is a valid \code{dim}
argument for an array in which all dimensions
have length greater than 0.
}
\examples{
chk_positive_dim(x = c(2L, 1L, 3L),
                 name = "x")
}
