% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R
\name{chk_positive_length}
\alias{chk_positive_length}
\alias{err_positive_length}
\title{Check that a vector has positive length,
ie not length 0}
\usage{
chk_positive_length(x, name)

err_positive_length(x, name)
}
\arguments{
\item{x}{A vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a vector has positive length,
ie not length 0
}
\examples{
x <- 1.0
chk_positive_length(x, name = "x")
}
