% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_strictly_increasing}
\alias{chk_strictly_increasing}
\alias{err_strictly_increasing}
\title{Check that elements in a vector of
numbers or dates are strictly increasing}
\usage{
chk_strictly_increasing(x, name)

err_strictly_increasing(x, name)
}
\arguments{
\item{x}{A numeric vector}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that elements in a vector of
numbers or dates are strictly increasing
}
\examples{
x <- 1:5
chk_strictly_increasing(x, name = "x")
}
