% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R
\name{chk_trans_list}
\alias{chk_trans_list}
\title{Check that a list describing general transitions between
states is valid}
\usage{
chk_trans_list(x, name)
}
\arguments{
\item{x}{A list if character vectors.}

\item{name}{The name for \code{x} to be used
in error messages.}
}
\description{
Check that a list describing general transitions between
states is valid
}
\examples{
x <- list(S = "I",
          I = c("R", "D"))
chk_trans_list(x, name = "x")
}
