% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_valid_label_single_month}
\alias{chk_valid_label_single_month}
\alias{chk_valid_label_single_quarter}
\alias{err_valid_label_single_month}
\alias{err_valid_label_single_quarter}
\title{Check that a label has the correct format for a
single quarter or month}
\usage{
chk_valid_label_single_month(x, name)

chk_valid_label_single_quarter(x, name)

err_valid_label_single_month(x, name)

err_valid_label_single_quarter(x, name)
}
\arguments{
\item{x}{A string.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that a label has the correct format for a
single quarter or month
}
\examples{
chk_valid_label_single_month("2001 Jan", name = "x")
chk_valid_label_single_quarter("2001 Q1", name = "x")
}
