% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_valid_quantile}
\alias{chk_valid_quantile}
\alias{chk_quantiles_increasing}
\alias{err_valid_quantile}
\alias{err_quantiles_increasing}
\title{Check that a vector is a valid quantile}
\usage{
chk_valid_quantile(x, name)

chk_quantiles_increasing(x, name)

err_valid_quantile(x, name)

err_quantiles_increasing(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
\code{chk_valid_quantile} checks that the elements
of \code{x} conform to R conventions about
formatting quantiles. \code{chk_quantile_increasing}
assumes that \code{x} is correctly formatted,
and checks that the elements are strictly increasing.
}
\examples{
x <- c("2.5\%", "50\%", "97.5")
chk_valid_quantile(x, name = "x")
chk_quantile_increasing(x, name = "x")
}
