% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R, R/err-values.R
\name{chk_values_categories}
\alias{chk_values_categories}
\alias{err_values_categories}
\title{Check values for Categories labels}
\usage{
chk_values_categories(x, name)

err_values_categories(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is character
  \item at most one NA
  \item no blanks (zero-length strings)
  \item no duplicates
}
}
\details{
Function also called by \code{\link{chk_values_triangles}},
\code{\link{chk_values_directions}}, and
\code{\link{chk_values_quantiles}}.
}
\examples{
x <- c("Thailand", "Laos", "Cambodia")
chk_values_categories(x = x, name = "x")
}
