% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R
\name{chk_values_datepoints}
\alias{chk_values_datepoints}
\title{Check values for DatePoints labels}
\usage{
chk_values_datepoints(x, name)
}
\arguments{
\item{x}{A vector of class \code{\link[base]{Date}}.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is Date vector
  \item no duplicates
}
}
\examples{
x <- as.Date(c("2000-05-01", NA, "2000-02-01"))
chk_values_datepoints(x = x, name = "x")
}
