% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R, R/err-values.R
\name{chk_values_dateranges}
\alias{chk_values_dateranges}
\alias{err_values_dateranges}
\title{Check values for DateRanges labels}
\usage{
chk_values_dateranges(x, name)

err_values_dateranges(x, name)
}
\arguments{
\item{x}{A list of Date vectors, 
each of which has length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is list of date vectors of length 2
  \item second element is later than first (if neither NA)
  \item intervals do not overlap
}
}
\examples{
x <- list(as.Date(c("2000-01-03", "2000-01-03")),
          as.Date(c(NA, "2000-01-02")),
          as.Date(c(NA, NA)))
chk_values_dateranges(x = x, name = "x")
}
