% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R, R/err-values.R
\name{chk_values_directions}
\alias{chk_values_directions}
\alias{err_values_directions}
\title{Check values for Directions labels}
\usage{
chk_values_directions(x, name)

err_values_directions(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is character
  \item no duplicates
  \item values belong to set {NA, "In", "Out"}
}
}
\examples{
x <- c("In", "Out")
chk_values_directions(x = x, name = "x")
}
