% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R
\name{chk_values_integers}
\alias{chk_values_integers}
\title{Check values for Integer labels}
\usage{
chk_values_integers(x, name)
}
\arguments{
\item{x}{A list of integer vectors, 
each of which has length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Specialised version of Intervals labels.
}
\details{
Checks:
\itemize{
  \item is list of integer vectors of length 2
  \item second element of each vector one greater than first
  \item intervals do not overlap
}
}
\examples{
x <- list(c(100L, NA), c(0L, 1L), c(NA, 0L), c(5L, 6L))
chk_values_integers(x = x, name = "x")
}
