% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R, R/err-values.R
\name{chk_values_intervals}
\alias{chk_values_intervals}
\alias{err_values_intervals}
\title{Check values for Intervals labels}
\usage{
chk_values_intervals(x, name)

err_values_intervals(x, name)
}
\arguments{
\item{x}{A list of integer vectors, 
each of which has length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is list of integer vectors of length 2
  \item second element of each vector greater than first
  \item intervals do not overlap
}
}
\details{
Identical to \code{\link{chk_values_quantities}}
}
\examples{
x <- list(c(100L, NA), c(0L, 5L), c(NA, 0L), c(5L, 50L))
chk_values_intervals(x = x, name = "x")
}
