% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R
\name{chk_values_months}
\alias{chk_values_months}
\title{Check values for Months labels}
\usage{
chk_values_months(x, name)
}
\arguments{
\item{x}{A list of Date vectors, 
each of which has length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is list of date vectors of length 2
  \item both elements are first day of month
  \item second element is one month later than first (if neither NA)
  \item intervals do not overlap
}
}
\examples{
x <- list(as.Date(c("2000-01-01", "2000-02-01")),
          as.Date(c(NA, "2000-01-01")),
          as.Date(c(NA, NA)))
chk_values_months(x = x, name = "x")
}
