% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R, R/err-values.R
\name{chk_values_pairs}
\alias{chk_values_pairs}
\alias{err_values_pairs}
\title{Checking function for subclasses of Pairs virtual superclass}
\usage{
chk_values_pairs(x, name)

err_values_pairs(x, name)
}
\arguments{
\item{x}{A list of vectors, 
each of which has length 2.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Function called by \code{\link{chk_values_integers}},
\code{\link{chk_values_intervals}},
\code{\link{chk_values_quantities}},
\code{\link{chk_values_quarters}},
\code{\link{chk_values_months}}, and
\code{\link{chk_values_dateranges}},
}
\details{
Checks:
\itemize{
  \item is list of vectors of length 2
  \item second element of each vector greater than first
  \item intervals do not overlap (including NAs)
}
}
\examples{
x <- list(c(100L, NA), c(0L, 5L), c(NA, 0L), c(5L, 50L))
chk_values_pairs(x = x, name = "x")
x <- list(as.Date(c("2001-01-01", "2002-01-01")))
chk_values_pairs(x = x, name = "x")
}
