% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-values.R
\name{chk_values_triangles}
\alias{chk_values_triangles}
\title{Check values for Triangles labels}
\usage{
chk_values_triangles(x, name)
}
\arguments{
\item{x}{A character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Checks:
\itemize{
  \item is character
  \item no duplicates
  \item values belong to set {NA, "Lower", "Upper"}
}
}
\examples{
x <- c("Upper", "Lower")
chk_values_triangles(x = x, name = "x")
}
