% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_zero}
\alias{chk_zero}
\alias{chk_zero_if_zero}
\alias{chk_zero_onlyif_zero}
\alias{chk_zero_ifonlyif_zero}
\alias{err_zero_if_zero}
\alias{err_zero_onlyif_zero}
\alias{err_zero_ifonlyif_zero}
\title{Check if 'x1' is zero, conditional on whether 'x2' is zero}
\usage{
chk_zero_if_zero(x1, x2, name1, name2)

chk_zero_onlyif_zero(x1, x2, name1, name2)

chk_zero_ifonlyif_zero(x1, x2, name1, name2)

err_zero_if_zero(x1, x2, name1, name2)

err_zero_onlyif_zero(x1, x2, name1, name2)

err_zero_ifonlyif_zero(x1, x2, name1, name2)
}
\arguments{
\item{x1}{An argument that could be \code{0}.}

\item{x2}{An argument that could be \code{0}.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
\code{chk_zero_if_zero} checks that \code{x1} is \code{0},
given that \code{x2} is \code{0}.
}
\details{
\code{chk_zero_onlyif_zero} checks that \code{x1} is only
\code{0} when \code{x2} is \code{0}.

\code{chk_zero_ifonlyif_zero} checks that \code{x1} and
\code{x2} are both \code{0} or are both non-\code{0}.
}
\examples{
x1 <- 0
x2 <- 2
chk_zero_if_zero(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 0
x2 <- 0
chk_zero_if_zero(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 1
x2 <- 0
chk_zero_onlyif_zero(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 0
x2 <- 0
chk_zero_onlyif_zero(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 1
x2 <- 2
chk_zero_ifonlyif_zero(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
x1 <- 0
x2 <- 0
chk_zero_ifonlyif_zero(x1 = x1, x2 = x2, name1 = "x1", name2 = "x2")
}
\seealso{
\code{\link{chk_zero}}
}
