% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{composite}
\alias{composite}
\alias{chk_all_class}
\alias{chk_array_metadata_complete}
\alias{chk_character_complete}
\alias{chk_dimnames_complete}
\alias{chk_ge_age_min}
\alias{chk_ge_year_min}
\alias{chk_is_first_day_unit}
\alias{chk_is_first_day_unit_consec}
\alias{chk_is_ge_scalar}
\alias{chk_is_ge_vector}
\alias{chk_is_gt_scalar}
\alias{chk_is_gt_vector}
\alias{chk_is_integer_consec}
\alias{chk_is_logical_flag}
\alias{chk_is_multiple_of}
\alias{chk_is_multiple_of_n}
\alias{chk_is_string}
\alias{chk_is_strictly_increasing}
\alias{chk_length_same}
\alias{chk_length_same_or_1}
\alias{chk_lt_age_max}
\alias{chk_lt_year_max}
\alias{chk_names_complete}
\alias{chk_names_dimnames_complete}
\alias{chk_trans_list}
\alias{err_all_class}
\alias{err_array_metadata_complete}
\alias{err_character_complete}
\alias{err_dimnames_complete}
\alias{err_ge_age_min}
\alias{err_ge_year_min}
\alias{err_is_first_day_unit}
\alias{err_is_first_day_unit_consec}
\alias{err_is_ge_scalar}
\alias{err_is_ge_vector}
\alias{err_is_gt_scalar}
\alias{err_is_gt_vector}
\alias{err_is_integer_consec}
\alias{err_is_logical_flag}
\alias{err_is_multiple_of}
\alias{err_is_multiple_of_n}
\alias{err_is_non_negative_scalar}
\alias{err_is_non_negative_vector}
\alias{err_is_positive_scalar}
\alias{err_is_positive_vector}
\alias{err_is_string}
\alias{err_is_strictly_increasing}
\alias{err_length_same}
\alias{err_length_same_or_1}
\alias{err_lt_age_max}
\alias{err_lt_year_max}
\alias{err_names_complete}
\alias{err_names_dimnames_complete}
\alias{err_trans_list}
\title{Checks of composite conditions}
\usage{
chk_all_class(x, name, class)

chk_array_metadata_complete(x, name)

chk_character_complete(x, name)

chk_dimnames_complete(x, name)

chk_ge_age_min(age, age_min, date, dob, unit)

chk_ge_year_min(date, year_min, first_month)

chk_is_first_day_unit(x, name, unit)

chk_is_first_day_unit_consec(x, name, unit)

chk_is_ge_scalar(x1, x2, name1, name2)

chk_is_ge_vector(x1, x2, name1, name2)

chk_is_gt_scalar(x1, x2, name1, name2)

chk_is_gt_vector(x1, x2, name1, name2)

chk_is_integer_consec(x, name)

chk_is_logical_flag(x, name)

chk_is_multiple_of(x1, x2, name1, name2, inf_ok)

chk_is_multiple_of_n(x, name, n, inf_ok)

chk_is_string(x, name)

chk_is_strictly_increasing(x, name)

chk_length_same(x1, x2, name1, name2)

chk_length_same_or_1(x1, x2, name1, name2)

chk_lt_age_max(age, age_max, date, dob, unit)

chk_lt_year_max(date, year_max, first_month)

chk_names_complete(x, name)

chk_names_dimnames_complete(x, name)

chk_trans_list(x, name)

err_all_class(x, name, class)

err_array_metadata_complete(x, name)

err_character_complete(x, name)

err_dimnames_complete(x, name)

err_ge_age_min(age, age_min, date, dob, unit)

err_ge_year_min(date, year_min, first_month)

err_is_first_day_unit(x, name, unit)

err_is_first_day_unit_consec(x, name, unit)

err_is_ge_scalar(x1, x2, name1, name2)

err_is_ge_vector(x1, x2, name1, name2)

err_is_gt_scalar(x1, x2, name1, name2)

err_is_gt_vector(x1, x2, name1, name2)

err_is_integer_consec(x, name)

err_is_logical_flag(x, name)

err_is_multiple_of(x1, x2, name1, name2, inf_ok)

err_is_multiple_of_n(x, name, n, inf_ok)

err_is_non_negative_scalar(x, name)

err_is_non_negative_vector(x, name)

err_is_positive_scalar(x, name)

err_is_positive_vector(x, name)

err_is_string(x, name)

err_is_strictly_increasing(x, name)

err_length_same(x1, x2, name1, name2)

err_length_same_or_1(x1, x2, name1, name2)

err_lt_age_max(age, age_max, date, dob, unit)

err_lt_year_max(date, year_max, first_month)

err_names_complete(x, name)

err_names_dimnames_complete(x, name)

err_trans_list(x, name)
}
\arguments{
\item{x}{The object being checked.}

\item{name}{The name used in any message. Typically,
but not always, the name of \code{x}.}

\item{class}{Name of class.}

\item{age}{Age, typically in years, but can be other unit.}

\item{age_min}{Minimum age. All ages must be greater than
or equal to this age. (Unlike with \code{age_max}, equality
is permitted.)}

\item{date}{Date on which event occurred or measurement made.
Object of class "Date".}

\item{dob}{Date of birth. Object of class "Date".}

\item{unit}{Measurement units for time, eg \code{"month"}.}

\item{year_min}{Minimum year. All dates must be greater than
or equal to the date implied by \code{year_min} and
\code{first_month}.}

\item{first_month}{A abbreviated month name. Used for defining
year-long periods.}

\item{x1}{The first of a pair of objects being checked.}

\item{x2}{The second of a pair of objects being checked.}

\item{name1}{The name of the first of the pair of objects.}

\item{name2}{The name of the second of the pair of objects.}

\item{inf_ok}{Whether to allow \code{Inf} (or sometimes
\code{-Inf}.)}

\item{n}{An integer.}

\item{age_max}{Maximum age. All ages must be less than this
age. (Unlike with \code{age_min}, equality is not permitted.)}

\item{year_max}{Maximum year. All dates must be less than
the date implied by \code{year_max} and \code{first_month}.}
}
\value{
When \code{x} passes the test,
the \code{chk*} and \code{err*} functions both
return \code{TRUE}.  When \code{x} fails the test,
the \code{chk*} functions return a string, and the \code{err*}
functions raise an error.
}
\description{
Functions to check for conditions that involve
several attributes. For instance, \code{chk_is_integer_consec}
checks whether \code{x} has type integer, whether \code{x}
has no \code{NA}s, and whether \code{all(diff(x)) == 1L}.
}
\details{
\code{chk_is_first_day_unit} checks whether a \code{x}
consists of first days for the time unit supplied,
eg values such \code{"2001-03-01"}, \code{"2001-05-01"}, or
\code{"2000-09-01"} when \code{unit} is \code{"months"}.
\code{chk_is_first_day_unit_consec} adds the condition
that these dates be consecutive, eg
\code{"2001-03-01"}, \code{"2001-04-01"}, \code{"2000-09-01"}.
Separately checking for first days makes the job of
\code{chk_is_first_day_unit_consec} much easier, which
is why we have two functions.

Most of the functions raise an error if \code{x} has \code{NA}s.
}
\seealso{
\code{\link{single}}, \code{\link{member}}
}
