% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy}
\alias{err_tdy}
\alias{err_tdy_breaks}
\alias{err_tdy_date}
\alias{err_tdy_min_max}
\alias{err_tdy_unit}
\title{Try to Tidy a Value}
\usage{
err_tdy_breaks(x, name)

err_tdy_date(x, name)

err_tdy_min_max(x1, x2, name1, name2)

err_tdy_unit(x, name)
}
\arguments{
\item{x}{The object being tidied.}

\item{name}{The name used in the error message. Typically,
but not always, the name of \code{x}.}

\item{x1}{The first of a pair of objects being checked.}

\item{x2}{The second of a pair of objects being checked.}

\item{name1}{The name of the first of the pair of objects.}

\item{name2}{The name of the second of the pair of objects.}
}
\value{
When err_tdy* can format \code{x} as required,
it returns the value; otherwise it raises an error.
}
\description{
Functions that try to put a value into a particular
format, and throw an error if this cannot be done.
These function do not have an chk* version, since
the results would be ambiguous in cases where the
return value was a string. Also, argument-tidying
is almost always done as part of the user interface,
when it is appropriate to raise an error if a value
departs from expectations.
}
\details{
In function \code{err_tdy_unit}, \code{x}
can be one of \code{NULL}, \code{"year"},
\code{"quarter"}, \code{"month"},
\code{"years"}, \code{"quarters"}, or \code{"months"}.
It can also have the form \code{"n year"}, or \code{"n years"},
where \code{n} is a positive integer.
\code{NULL} is equivalent to \code{"year"}.
}
