% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_break_min_max_date}
\alias{err_tdy_break_min_max_date}
\title{Check and tidy minimum and maximum dates}
\usage{
err_tdy_break_min_max_date(break_min, break_max, unit, null_ok, equal_ok)
}
\arguments{
\item{break_min}{Date or NULL. If non-NULL, lowest permissible value
for \code{date}.}

\item{break_max}{Date or NULL. If non-NULL, highest permissible value
for \code{date}.}

\item{unit}{A time unit.}

\item{null_ok}{Whether \code{break_min} and \code{break_max} can be
\code{NULL}}

\item{equal_ok}{Whether \code{break_min} and \code{break_max} can be
\code{NULL}.}
}
\description{
Check and tidy minimum and maximum dates
}
\examples{
err_tdy_break_min_max_date(break_min = as.Date("2000-01-01"),
                           break_max = as.Date("2080-01-01"),
                           unit = "year",
                           null_ok = FALSE,
                           equal_ok = FALSE)
}
\seealso{
\code{\link{err_tdy_break_min_max_integer}}
}
