% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_break_min_max_integer}
\alias{err_tdy_break_min_max_integer}
\title{Check and tidy minimum and maximum integers}
\usage{
err_tdy_break_min_max_integer(break_min, break_max, null_ok, equal_ok)
}
\arguments{
\item{break_min}{An integer, a number that can
be coercted to integer, or NULL.}

\item{break_max}{An integer, a number that can
be coercted to integer, or NULL.}

\item{null_ok}{Whether \code{break_min} and \code{break_max} can be
\code{NULL}}

\item{equal_ok}{Whether \code{break_min} and \code{break_max} can be
\code{NULL}.}
}
\description{
Typically used for age or cohort.
}
\examples{
err_tdy_break_min_max_integer(break_min = 0L,
                              break_max = 100L,
                              null_ok = FALSE,
                              equal_ok = FALSE)
}
\seealso{
\code{\link{err_tdy_break_min_max_date}},
}
