% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_breaks_date_cohort}
\alias{err_tdy_breaks_date_cohort}
\title{Check and tidy a vector of breaks used to define cohorts}
\usage{
err_tdy_breaks_date_cohort(breaks, open_first)
}
\arguments{
\item{breaks}{A vector of dates, or values that can be coerced to dates.}

\item{open_first}{Whether the first interval is open on the left.}
}
\description{
Check and tidy a vector of breaks used to define cohorts
}
\examples{
breaks <- c("2000-01-01", "2010-01-01", "2020-01-01")
err_tdy_breaks_date_cohort(breaks = breaks,
                           open_first = TRUE)
}
\seealso{
\code{\link{err_tdy_breaks_date_period}},
\code{\link{err_tdy_breaks_integer_age}}
\code{\link{err_tdy_breaks_integer_enum}}
}
