% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_breaks_date_period}
\alias{err_tdy_breaks_date_period}
\title{Check and tidy a vector of breaks used to define periods}
\usage{
err_tdy_breaks_date_period(breaks)
}
\arguments{
\item{breaks}{A vector of dates, or values that can be coerced to dates.}
}
\description{
Check and tidy a vector of breaks used to define periods
}
\examples{
breaks <- c("2000-01-01", "2010-01-01", "2020-01-01")
err_tdy_breaks_date_period(breaks)
}
\seealso{
\code{\link{err_tdy_breaks_date_cohort}},
\code{\link{err_tdy_breaks_integer_age}}
\code{\link{err_tdy_breaks_integer_enum}}
}
