% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_breaks_integer_age}
\alias{err_tdy_breaks_integer_age}
\title{Check and tidy a vector of integer breaks used to define age groups}
\usage{
err_tdy_breaks_integer_age(breaks, open_last)
}
\arguments{
\item{breaks}{An integer vector.}

\item{open_last}{Whether the final interval is open on the right.}
}
\description{
Check and tidy a vector of integer breaks used to define age groups
}
\examples{
breaks <- c(0L, 15L, 65L)
err_tdy_breaks_integer_age(breaks,
                           open_last = TRUE)
}
\seealso{
\code{\link{err_tdy_breaks_date_cohort}},
\code{\link{err_tdy_breaks_date_period}},
\code{\link{err_tdy_breaks_integer_enum}}
}
