% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_breaks_integer_cohort}
\alias{err_tdy_breaks_integer_cohort}
\title{Check and tidy a vector of integer breaks used to define periods}
\usage{
err_tdy_breaks_integer_cohort(breaks, open_first)
}
\arguments{
\item{breaks}{An integer vector.}

\item{open_first}{Whether the first interval is open on the left.}
}
\description{
Check and tidy a vector of integer breaks used to define periods
}
\examples{
breaks <- c(2000, 2010, 2020)
err_tdy_breaks_integer_cohort(breaks,
                              open_first = TRUE)
}
\seealso{
\code{\link{err_tdy_breaks_date_cohort}},
\code{\link{err_tdy_breaks_date_period}},
\code{\link{err_tdy_breaks_integer_age}},
\code{\link{err_tdy_lower_upper_enumeration}},
\code{\link{err_tdy_breaks_integer_period}}
}
