% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_breaks_integer_enum}
\alias{err_tdy_breaks_integer_enum}
\title{Check and tidy a vector of breaks used to define enumerations}
\usage{
err_tdy_breaks_integer_enum(breaks, open_first, open_last)
}
\arguments{
\item{breaks}{An integer vector.}

\item{open_first}{Whether the first interval is open on the left.}

\item{open_last}{Whether the final interval is open on the right.}
}
\description{
Check and tidy a vector of breaks used to define enumerations
}
\examples{
breaks <- c(0, 1000, 5000, 20000)
err_tdy_breaks_integer_enum(breaks,
                            open_first = TRUE,
                            open_last = TRUE)
}
\seealso{
\code{\link{err_tdy_breaks_date_cohort}},
\code{\link{err_tdy_breaks_date_period}},
\code{\link{err_tdy_breaks_integer_age}}
}
