% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_breaks_integer_period}
\alias{err_tdy_breaks_integer_period}
\title{Check and tidy a vector of integer breaks used to define periods}
\usage{
err_tdy_breaks_integer_period(breaks)
}
\arguments{
\item{breaks}{An integer vector.}
}
\description{
Check and tidy a vector of integer breaks used to define periods
}
\examples{
breaks <- c(2000, 2010, 2020)
err_tdy_breaks_integer_period(breaks)
}
\seealso{
\code{\link{err_tdy_breaks_date_cohort}},
\code{\link{err_tdy_breaks_date_period}},
\code{\link{err_tdy_breaks_integer_age}},
\code{\link{err_tdy_breaks_integer_cohort}},
\code{\link{err_tdy_breaks_integer_enum}}
}
