% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_integer_scalar}
\alias{err_tdy_integer_scalar}
\alias{err_tdy_integer_vector}
\title{Check and tidy integers or objects that can be coerced to integer}
\usage{
err_tdy_integer_scalar(x, name, null_ok = FALSE)

err_tdy_integer_vector(x, name)
}
\arguments{
\item{x}{A scalar or vector of integers, or quantities
that can be coerced to integer}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{null_ok}{Whether to allow \code{x} to be \code{NULL}.}
}
\description{
Check and tidy integers or objects that can be coerced to integer
}
\examples{
x <- 1.0
err_tdy_integer_scalar(x = x, name = "x", null_ok = FALSE)
x <- NULL
err_tdy_integer_scalar(x = x, name = "x", null_ok = TRUE)
x <- 3:1
err_tdy_integer_vector(x = x, name = "x")
}
\seealso{
\code{\link{err_tdy_date_scalar}}
}
