% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_lower_upper_enumeration}
\alias{err_tdy_lower_upper_enumeration}
\title{Check and tidy 'lower' and 'upper' vectors used to
define enumerations}
\usage{
err_tdy_lower_upper_enumeration(lower, upper)
}
\arguments{
\item{lower}{A vector that can be coerced to integer.}

\item{upper}{A vector that can be coerced to integer.}
}
\description{
The first element of 'lower' can be \code{NA},
implying that the first interval is open.
The last element of 'upper' can be \code{NA},
implying that the last interval is open.
}
\details{
There can be gaps between intervals, but intervals
must not overlap. The width of every interval
must be at least 1.
}
\examples{
lower <- c(NA, 0, 5, 20)
upper <- c(0, 5, 10, NA)
err_tdy_lower_upper_enumeration(lower = lower,
                                upper = upper)
}
\seealso{
\code{\link{err_tdy_breaks_date_cohort}},
\code{\link{err_tdy_breaks_date_period}},
\code{\link{err_tdy_breaks_integer_age}},
\code{\link{err_tdy_breaks_integer_cohort}},
\code{\link{err_tdy_breaks_integer_period}}
}
