% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_many_to_one}
\alias{err_tdy_many_to_one}
\title{Check and tidy a data frame describing a many-to-one mapping}
\usage{
err_tdy_many_to_one(x, name)
}
\arguments{
\item{x}{A data frame.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check and tidy a data frame describing a many-to-one mapping
}
\examples{
x <- data.frame(from = 1:3,
                to = c("A", "B", "A"))
err_tdy_many_to_one(x, name = "x")
}
