% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_map_pos}
\alias{err_tdy_map_pos}
\title{Check and tidy 'map_pos'}
\usage{
err_tdy_map_pos(map_pos, dim_self, dim_oth, map_dim)
}
\arguments{
\item{map_pos}{A list of integer vectors.}

\item{dim_self}{The dimensions of \code{self}.}

\item{dim_oth}{The dimensions of \code{oth}.}

\item{map_dim}{A vector of non-negative integers.}
}
\value{
\code{map_pos}, with elements coerced to integer.
}
\description{
Check and tidy \code{map_pos}, a mapping between
positions along dimensions of array \code{self} and
positions along dimensions of array \code{oth}.
\code{map_pos} is a list the same length as
\code{dim_self}. Each element of \code{map_pos}
is a vector, showing, for each position on the dimension
of \code{self}, which postition on the dimension
of \code{oth} it maps on to. Multiple positions
on \code{self} can map on to the same position on \code{oth}.
If a position on \code{self} does not map to a position,
on \code{oth}, the value for the non-mapping position
is \code{0}. If a dimension of \code{self}
does not have a corresponding dimension in
\code{oth}, then corresponding element of \code{map_dim}
is vector of zeros.
}
\details{
\code{err_tdy_map_pos} assumes that \code{map_dim},
\code{dim_self}, and \code{dim_oth} are all valid.

\code{self} and \code{oth} are both assumed to have positive
length, ie no dimensions with length 0.
}
\examples{
map_pos <- list(c(1L, 2L, 3L, 4L), c(0L, 1L), c(1L, 2L, 2L))
map_dim <- c(1L, 3L, 2L)
dim_self <- c(4L, 2L, 3L)
dim_oth <- c(4L, 2L, 1L)
err_tdy_map_pos(map_pos = map_pos,
                map_dim = map_dim,
                dim_self = dim_self,
                dim_oth = dim_oth)

map_pos <- list(c(2L, 1L), c(1L, 1L), c(0L, 0L, 0L))
map_dim <- c(1L, 2L, 0L)
dim_self <- c(2L, 2L, 3L)
dim_oth <- c(2L, 1L)
err_tdy_map_pos(map_pos = map_pos,
                map_dim = map_dim,
                dim_self = dim_self,
                dim_oth = dim_oth)
}
\seealso{
\code{\link{err_tdy_map_dim}}
}
