% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_month}
\alias{err_tdy_month}
\title{Check and tidy numeric or character vector
of months}
\usage{
err_tdy_month(x, name)
}
\arguments{
\item{x}{A numeric or character vector.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Convert a numeric or character vector of month
indices or codes to an integer vector.
The vector can contain NAs.
}
\examples{
err_tdy_month(x = c(1, NA, 12, 5),
              name = "x")
err_tdy_month(x = c("01", NA, "12", "05"),
              name = "x")
err_tdy_month(x = c("Jan", NA, "Dec", "May"),
              name = "x")
err_tdy_month(x = c("January", NA, "December", "May"),
              name = "x")
}
