% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_month_label}
\alias{err_tdy_month_label}
\title{Check and tidy month label}
\usage{
err_tdy_month_label(x, name)
}
\arguments{
\item{x}{A month label.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that label valid, and return
date of first day of month.
Allows 0 or more spaces is case insensitive.
}
\examples{
err_tdy_month_label(x = "2000 Jan", name = "x")
err_tdy_month_label(x = "2000Jan", name = "x")
err_tdy_month_label(x = "2000 JAN", name = "x")
}
