% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_month_start}
\alias{err_tdy_month_start}
\title{Check and tidy 'month_start' argument,
used in defining years}
\usage{
err_tdy_month_start(x, name)
}
\arguments{
\item{x}{An abbreviated month name.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check and tidy 'month_start' argument,
used in defining years
}
\examples{
x <- "Feb"
err_tdy_month_start(x = x, name = "x")
}
\seealso{
\code{\link[demprep]{date_to_period_year}},
}
