% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_positive_integer}
\alias{err_tdy_positive_integer}
\alias{err_tdy_positive_integer_scalar}
\alias{err_tdy_positive_integer_vector}
\title{Check and tidy a non-negative integer scalar or vector}
\usage{
err_tdy_positive_integer_scalar(x, name)

err_tdy_positive_integer_vector(x, name)
}
\arguments{
\item{x}{An integer scalar or vector, or somthing that
can be coerced to one.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check and tidy a non-negative integer scalar or vector
}
\examples{
x <- 3.0
err_tdy_positive_integer_scalar(x = x, name = "x")
x <- c(3.0, 1.0, 4.0)
err_tdy_positive_integer_vector(x = x, name = "x")
}
\seealso{
\code{\link{err_tdy_non_negative_integer_scalar}}
}
