% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_positive_integer_scalar}
\alias{err_tdy_positive_integer_scalar}
\title{Check and tidy a non-negative integer scalar}
\usage{
err_tdy_positive_integer_scalar(x, name, null_ok = FALSE)
}
\arguments{
\item{x}{An integer, or somthing that
can be coerced to one.}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{null_ok}{Whether \code{x} can be \code{NULL}.}
}
\description{
Check and tidy a non-negative integer scalar
}
\examples{
x <- 3.0
err_tdy_positive_integer_scalar(x = x, name = "x")
}
\seealso{
\code{\link{err_tdy_non_negative_integer_scalar}}
}
