% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_quarter_label}
\alias{err_tdy_quarter_label}
\title{Check and tidy quarter label}
\usage{
err_tdy_quarter_label(x, name)
}
\arguments{
\item{x}{A quarter label.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that label valid, and return
date of first day of quarter.
Allows 0 or more spaces and lower-case
q.
}
\examples{
err_tdy_quarter_label(x = "2000 Q1", name = "x")
err_tdy_quarter_label(x = "2000Q1", name = "x")
err_tdy_quarter_label(x = "2000 q1", name = "x")
}
