% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy_same_length}
\alias{err_tdy_same_length}
\title{Ensure that two vectors have same length,
if necessary by replicating a length-1 vector}
\usage{
err_tdy_same_length(x1, x2, name1, name2)
}
\arguments{
\item{x1}{An object.}

\item{x2}{An object.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}
}
\description{
Ensure that two vectors have same length,
if necessary by replicating a length-1 vector
}
\examples{
x1 <- 1:5
x2 <- 1
err_tdy_same_length(x1 = x1, x2 = x2,
                    name1 = "x1", name2 = "x2")
}
\seealso{
\code{\link{chk_length_same_or_1}}
}
