% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-member.R, R/err-member.R
\name{member}
\alias{member}
\alias{chk_member_time_unit}
\alias{err_member_time_unit}
\title{Checks of Membership}
\usage{
chk_member_time_unit(x, name)

err_member_time_unit(x, name)
}
\arguments{
\item{x}{The object being checked.}

\item{name}{The name used in any message. Typically,
but not always, the name of \code{x}.}
}
\value{
When \code{x} passes the test,
the \code{chk*} and \code{err*} functions both
return \code{TRUE}.  When \code{x} fails the test,
the \code{chk*} functions return a string, and the \code{err*}
functions raise an error.
}
\description{
Functions to check whether a string belongs to a
set of values.  Includes a check of
whether \code{x} is a valid string, to provide
more helpful error messages to users.
}
\seealso{
\code{\link{single}}, \code{\link{composite}}
}
