% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-single.R, R/err-single.R, R/err_tdy.R
\name{single}
\alias{single}
\alias{chk_is_character}
\alias{chk_is_date}
\alias{chk_is_date_equiv}
\alias{chk_is_integer}
\alias{chk_is_integer_equiv}
\alias{chk_is_length_1}
\alias{chk_is_logical}
\alias{chk_is_not_na_list}
\alias{chk_is_not_na_scalar}
\alias{chk_is_not_na_vector}
\alias{chk_is_numeric}
\alias{err_is_character}
\alias{err_is_date}
\alias{err_is_date_equiv}
\alias{err_is_integer}
\alias{err_is_integer_equiv}
\alias{err_is_length_1}
\alias{err_is_logical}
\alias{err_is_not_na_list}
\alias{err_is_not_na_scalar}
\alias{err_is_not_na_vector}
\alias{err_is_numeric}
\alias{err_tdy_integer}
\title{Checks of Single Conditions}
\usage{
chk_is_character(x, name)

chk_is_date(x, name)

chk_is_date_equiv(x, name)

chk_is_integer(x, name)

chk_is_integer_equiv(x, name)

chk_is_length_1(x, name)

chk_is_logical(x, name)

chk_is_not_na_list(x, name)

chk_is_not_na_scalar(x, name)

chk_is_not_na_vector(x, name)

chk_is_numeric(x, name)

err_is_character(x, name)

err_is_date(x, name)

err_is_date_equiv(x, name)

err_is_integer(x, name)

err_is_integer_equiv(x, name)

err_is_length_1(x, name)

err_is_logical(x, name)

err_is_not_na_list(x, name)

err_is_not_na_scalar(x, name)

err_is_not_na_vector(x, name)

err_is_numeric(x, name)

err_tdy_integer(x, name)
}
\arguments{
\item{x}{The object being checked.}

\item{name}{The name used in any message. Typically,
but not always, the name of \code{x}.}
}
\value{
When \code{x} passes the test,
the \code{chk*} and \code{err*} functions both
return \code{TRUE}.  When \code{x} fails the test,
the \code{chk*} functions return a string, and the \code{err*}
functions raise an error.
}
\description{
Functions to check single conditions.
No sanity-checking of arguments is
carried out. These functions are typically
called from  other, more complicated,
checking functions.
}
\seealso{
\code{\link{member}}, \code{\link{composite}}
}
