% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-specific.R, R/err-specific.R
\name{chk_map_dim}
\alias{chk_map_dim}
\alias{err_map_dim}
\title{Check 'map_dim'}
\usage{
chk_map_dim(x, name)

err_map_dim(x, name)
}
\arguments{
\item{x}{A vector of non-negative integers,
unique apart from any zeros.}

\item{name}{The name for \code{x} to be used
in error messages.}
}
\description{
Validity tests for a \code{\link[=err_tdy_map_dim]{map_dim}}
that only require the object itself.
}
\examples{
chk_map_dim(x = c(3L, 1L, 0L),
            name = "map_dim")
}
