% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_multiple_of}
\alias{chk_multiple_of}
\alias{err_multiple_of}
\title{Check that 'x1' is a multiple of 'x2'}
\usage{
chk_multiple_of(x1, x2, name1, name2, null_ok)

err_multiple_of(x1, x2, name1, name2, null_ok)
}
\arguments{
\item{x1}{A scalar.}

\item{x2}{A scalar.}

\item{name1}{The name for \code{x1} that
will be used in error messages.}

\item{name2}{The name for \code{x2} that
will be used in error messages.}

\item{null_ok}{Whether \code{x1} can be \code{NULL},}
}
\description{
If \code{null_ok} is \code{TRUE} and \code{x1}
is \code{NULL} then the test succeeds.
}
\examples{
x1 <- 10L
x2 <- 2L
chk_multiple_of(x1 = x1, x2 = x2,
                name1 = "x1", name2 = "x2",
                null_ok = FALSE)
x1 <- NULL
chk_multiple_of(x1 = x1, x2 = x2,
                name1 = "x1", name2 = "x2",
                null_ok = TRUE)
}
\seealso{
\code{\link{chk_multiple_of_n}}
}
