% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_multiple_of_n}
\alias{chk_multiple_of_n}
\alias{err_multiple_of_n}
\title{Check that 'x1' is a multiple of 'n'}
\usage{
chk_multiple_of_n(x, name, n, null_ok)

err_multiple_of_n(x, name, n, null_ok)
}
\arguments{
\item{x}{A scalar.}

\item{name}{The name for \code{x} that
will be used in error messages.}

\item{n}{A scalar.}

\item{null_ok}{Whether \code{x} can be \code{NULL},}
}
\description{
\code{chk_multiple_of_n} differs from
\code{\link{chk_multiple_of}} only in the
error message. \code{chk_multiple_of_n} refers
to the value of \code{n}, rather than to the
variable \code{"n"}.
}
\examples{
x1 <- 10L
n <- 2L
chk_multiple_of_n(x1 = x1, n = n, name1 = "x",
                  null_ok = FALSE)
x1 <- NULL
chk_multiple_of(x1 = x1, n = n, name1 = "x",
                null_ok = FALSE)
}
