% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{chk_pos_initial}
\alias{chk_pos_initial}
\alias{err_pos_initial}
\title{Check that an initial 'pos' argument consists
entirely of 1s}
\usage{
chk_pos_initial(x, name)

err_pos_initial(x, name)
}
\arguments{
\item{x}{An integer vector with positive length,
consisting entirely of 1s.}

\item{name}{The name for \code{x} that
will be used in error messages.}
}
\description{
Check that an initial 'pos' argument consists
entirely of 1s
}
\examples{
chk_pos_initial(x = c(1L, 1L, 1L),
                name = "pos")
}
\seealso{
\code{\link{chk_positive_dim}}
}
