% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{composite}
\alias{composite}
\alias{chk_is_first_day_time_unit}
\alias{chk_is_first_day_time_unit_consec}
\alias{chk_is_integer_consec}
\alias{chk_is_string}
\alias{chk_is_strictly_increasing}
\alias{err_is_first_day_time_unit}
\alias{err_is_first_day_time_unit_consec}
\alias{err_is_integer_consec}
\alias{err_is_string}
\alias{err_is_strictly_increasing}
\title{Checks of Composite Conditions}
\usage{
chk_is_first_day_time_unit(x, name, time_unit)

chk_is_first_day_time_unit_consec(x, name, time_unit)

chk_is_integer_consec(x, name)

chk_is_string(x, name)

chk_is_strictly_increasing(x, name)

err_is_first_day_time_unit(x, name, time_unit)

err_is_first_day_time_unit_consec(x, name, time_unit)

err_is_integer_consec(x, name)

err_is_string(x, name)

err_is_strictly_increasing(x, name)
}
\arguments{
\item{x}{The object being checked.}

\item{name}{The name used in any message. Typically,
but not always, the name of \code{x}.}

\item{time_unit}{Measurement units for time, eg \code{"month"}.}
}
\value{
When \code{x} passes the test,
the \code{chk*} and \code{err*} functions both
return \code{TRUE}.  When \code{x} fails the test,
the \code{chk*} functions return a string, and the \code{err*}
functions raise an error.
}
\description{
Functions to check for conditions that involve
several attributes. For instance, \code{chk_is_integer_consec}
checks whether \code{x} has type integer, whether \code{x}
has no \code{NA}s, and whether \code{all(diff(x)) == 1L}.
}
\seealso{
\code{\link{single}}, \code{\link{member}}
}
