% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-composite.R, R/err-composite.R
\name{composite}
\alias{composite}
\alias{chk_all_class}
\alias{chk_array_metadata_complete}
\alias{chk_character_complete}
\alias{chk_date_consistent_with_month_start}
\alias{chk_date_consistent_with_width_origin}
\alias{chk_dimnames_complete}
\alias{chk_labels_valid_for_dimtype}
\alias{chk_ge_break_min_age}
\alias{chk_ge_break_min_date}
\alias{chk_is_first_day_unit_scalar}
\alias{chk_is_first_day_unit_vector}
\alias{chk_is_first_day_unit_consec}
\alias{chk_is_ge_scalar}
\alias{chk_is_ge_vector}
\alias{chk_is_gt_scalar}
\alias{chk_is_gt_vector}
\alias{chk_is_integer_consec}
\alias{chk_is_logical_flag}
\alias{chk_is_multiple_of}
\alias{chk_is_multiple_of_n}
\alias{chk_is_string}
\alias{chk_is_strictly_increasing}
\alias{chk_is_valid_quantile}
\alias{chk_length_same}
\alias{chk_length_same_or_1}
\alias{chk_lt_break_max_age}
\alias{chk_lt_break_max_date}
\alias{chk_names_complete}
\alias{chk_names_dimnames_complete}
\alias{chk_trans_list}
\alias{err_all_class}
\alias{err_array_metadata_complete}
\alias{err_character_complete}
\alias{err_date_consistent_with_month_start}
\alias{err_date_consistent_with_width_origin}
\alias{err_dimnames_complete}
\alias{err_labels_valid_for_dimtype}
\alias{err_ge_break_min_age}
\alias{err_ge_break_min_date}
\alias{err_is_first_day_unit_scalar}
\alias{err_is_first_day_unit_vector}
\alias{err_is_first_day_unit_consec}
\alias{err_is_ge_scalar}
\alias{err_is_ge_vector}
\alias{err_is_gt_scalar}
\alias{err_is_gt_vector}
\alias{err_is_integer_consec}
\alias{err_is_logical_flag}
\alias{err_is_multiple_of}
\alias{err_is_multiple_of_n}
\alias{err_is_non_negative_scalar}
\alias{err_is_non_negative_vector}
\alias{err_is_positive_scalar}
\alias{err_is_positive_vector}
\alias{err_is_string}
\alias{err_is_strictly_increasing}
\alias{err_is_valid_quantile}
\alias{err_length_same}
\alias{err_length_same_or_1}
\alias{err_lt_break_max_age}
\alias{err_lt_break_max_date}
\alias{err_names_complete}
\alias{err_names_dimnames_complete}
\alias{err_trans_list}
\title{Checks of composite conditions}
\usage{
chk_all_class(x, name, class)

chk_array_metadata_complete(x, name)

chk_character_complete(x, name)

chk_date_consistent_with_month_start(x, name, month_start)

chk_date_consistent_with_width_origin(x, name, origin, width)

chk_dimnames_complete(x, name)

chk_labels_valid_for_dimtype(labels, dimtype)

chk_ge_break_min_age(age, break_min, date, dob, unit)

chk_ge_break_min_date(date, break_min)

chk_is_first_day_unit_scalar(x, name, unit)

chk_is_first_day_unit_vector(x, name, unit)

chk_is_first_day_unit_consec(x, name, unit)

chk_is_ge_scalar(x1, x2, name1, name2)

chk_is_ge_vector(x1, x2, name1, name2)

chk_is_gt_scalar(x1, x2, name1, name2)

chk_is_gt_vector(x1, x2, name1, name2)

chk_is_integer_consec(x, name)

chk_is_logical_flag(x, name)

chk_is_multiple_of(x1, x2, name1, name2, null_ok)

chk_is_multiple_of_n(x, name, n, null_ok)

chk_is_string(x, name)

chk_is_strictly_increasing(x, name)

chk_is_valid_quantile(x, name)

chk_length_same(x1, x2, name1, name2)

chk_length_same_or_1(x1, x2, name1, name2)

chk_lt_break_max_age(age, break_max, date, dob, unit)

chk_lt_break_max_date(date, break_max)

chk_names_complete(x, name)

chk_names_dimnames_complete(x, name)

chk_trans_list(x, name)

err_all_class(x, name, class)

err_array_metadata_complete(x, name)

err_character_complete(x, name)

err_date_consistent_with_month_start(x, name, month_start)

err_date_consistent_with_width_origin(x, name, origin, width)

err_dimnames_complete(x, name)

err_labels_valid_for_dimtype(labels, dimtype)

err_ge_break_min_age(age, break_min, date, dob, unit)

err_ge_break_min_date(date, break_min)

err_is_first_day_unit_scalar(x, name, unit)

err_is_first_day_unit_vector(x, name, unit)

err_is_first_day_unit_consec(x, name, unit)

err_is_ge_scalar(x1, x2, name1, name2)

err_is_ge_vector(x1, x2, name1, name2)

err_is_gt_scalar(x1, x2, name1, name2)

err_is_gt_vector(x1, x2, name1, name2)

err_is_integer_consec(x, name)

err_is_logical_flag(x, name)

err_is_multiple_of(x1, x2, name1, name2, null_ok)

err_is_multiple_of_n(x, name, n, null_ok)

err_is_non_negative_scalar(x, name)

err_is_non_negative_vector(x, name)

err_is_positive_scalar(x, name)

err_is_positive_vector(x, name)

err_is_string(x, name)

err_is_strictly_increasing(x, name)

err_is_valid_quantile(x, name)

err_length_same(x1, x2, name1, name2)

err_length_same_or_1(x1, x2, name1, name2)

err_lt_break_max_age(age, break_max, date, dob, unit)

err_lt_break_max_date(date, break_max)

err_names_complete(x, name)

err_names_dimnames_complete(x, name)

err_trans_list(x, name)
}
\arguments{
\item{x}{The object being checked.}

\item{name}{The name used in any message. Typically,
but not always, the name of \code{x}.}

\item{class}{Name of class.}

\item{month_start}{An abbreviated month name,
as generated by \code{\link[base]{month.abb}}.}

\item{origin}{A year used to define multi-year periods.}

\item{width}{The width of multi-year age groups or periods.}

\item{labels}{Dimension labels: the \code{i}th element of \code{dimnames}}

\item{dimtype}{Character. The dimtype of the dimension.}

\item{age}{Age, typically in years, but can be other unit.}

\item{break_min}{First value for breaks.}

\item{date}{Date on which event occurred or measurement made.
Object of class "Date".}

\item{dob}{Date of birth. Object of class "Date".}

\item{unit}{Measurement units for time, eg \code{"month"}.}

\item{x1}{The first of a pair of objects being checked.}

\item{x2}{The second of a pair of objects being checked.}

\item{name1}{The name of the first of the pair of objects.}

\item{name2}{The name of the second of the pair of objects.}

\item{null_ok}{Whether to allow \code{NULL}.}

\item{n}{An integer.}

\item{break_max}{Last value for breaks.}
}
\value{
When \code{x} passes the test,
the \code{chk*} and \code{err*} functions both
return \code{TRUE}.  When \code{x} fails the test,
the \code{chk*} functions return a string, and the \code{err*}
functions raise an error.
}
\description{
Functions to check for conditions that involve
several attributes. For instance, \code{chk_is_integer_consec}
checks whether \code{x} has type integer, whether \code{x}
has no \code{NA}s, and whether \code{all(diff(x)) == 1L}.
}
\details{
\code{chk_is_first_day_unit} checks whether a \code{x}
consists of first days for the time unit supplied,
eg values such \code{"2001-03-01"}, \code{"2001-05-01"}, or
\code{"2000-09-01"} when \code{unit} is \code{"months"}.
\code{chk_is_first_day_unit_consec} adds the condition
that these dates be consecutive, eg
\code{"2001-03-01"}, \code{"2001-04-01"}, \code{"2000-09-01"}.
Separately checking for first days makes the job of
\code{chk_is_first_day_unit_consec} much easier, which
is why we have two functions.

Most of the functions raise an error if \code{x} has \code{NA}s.
}
\seealso{
\code{\link{single}}, \code{\link{member}}
}
