% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_tdy.R
\name{err_tdy}
\alias{err_tdy}
\alias{err_tdy_breaks_integer}
\alias{err_tdy_date}
\alias{err_tdy_date_dob}
\alias{err_tdy_first_month}
\alias{err_tdy_integer_scalar}
\alias{err_tdy_integer_vector}
\alias{err_tdy_many_to_one}
\alias{err_tdy_positive_integer_scalar}
\alias{err_tdy_same_length}
\alias{err_tdy_unit}
\title{Try to tidy a value}
\usage{
err_tdy_breaks_integer(x, name)

err_tdy_date(x, name)

err_tdy_date_dob(date, dob)

err_tdy_first_month(x, name)

err_tdy_integer_scalar(x, name, inf_ok = FALSE)

err_tdy_integer_vector(x, name)

err_tdy_many_to_one(x, name)

err_tdy_positive_integer_scalar(x, name, inf_ok = FALSE)

err_tdy_same_length(x1, x2, name1, name2)

err_tdy_unit(x, name)
}
\arguments{
\item{x}{The object being checked.}

\item{name}{The name used in any message. Typically,
but not always, the name of \code{x}.}

\item{date}{Date on which event occurred or measurement made.
Object of class "Date".}

\item{dob}{Date of birth. Object of class "Date".}

\item{inf_ok}{Whether to allow \code{Inf} (or sometimes
\code{-Inf}.) Defaults to \code{FALSE}.}

\item{x1}{The first of a pair of objects being checked.}

\item{x2}{The second of a pair of objects being checked.}

\item{name1}{The name of the first of the pair of objects.}

\item{name2}{The name of the second of the pair of objects.}
}
\value{
When err_tdy* can format \code{x} as required,
it returns the value; otherwise it raises an error.
}
\description{
Functions that try to put a value into a particular
format, and throw an error if this cannot be done.
These function do not have an chk* version, since
the results would be ambiguous in cases where the
return value was a string. Also, argument-tidying
is almost always done as part of the user interface,
when it is appropriate to raise an error if a value
departs from expectations.
}
\details{
In function \code{err_tdy_unit}, \code{x}
can be one of \code{NULL}, \code{"year"},
\code{"quarter"}, \code{"month"},
\code{"years"}, \code{"quarters"}, or \code{"months"}.
It can also have the form \code{"n year"}, or \code{"n years"},
where \code{n} is a positive integer.
\code{NULL} is equivalent to \code{"year"}.
}
